csharp doc, refactor Service.cs
