+++ RobotRaconteurNET/RobotRaconteurNET/MultiDimArray.cs
@@ -19,16 +19,32 @@ namespace RobotRaconteur
 {
 
 /// <summary>
-/// Represents an unpacked MultiDimArray.  This class stores
-/// the real and complex data as column vectors that is
-/// reshaped based on Dims.
+/// Numeric primitive multidimensional array value type
 /// </summary>
+/// <remarks>
+/// This class stores a numeric primitive multidimensional arrays.
+/// Multidimensional arrays are stored as a uint array of
+/// dimensions, and an array of the flattened elements.
+/// Arrays are stored in column major, or "Fortran" order.
+///
+/// Valid types for array are `bool`, `double`, `float`, `sbyte`, `byte`, `short`,
+/// `ushort`, `int`, `uint`, `long`, `ulong`, `CDouble`,
+/// or `CSingle`. Attempts to use any other types will result in a compiler error.
+/// </remarks>
 public class MultiDimArray
 {
 
+    /// <summary>
+    /// Construct empty MultiDimArray
+    /// </summary>
     public MultiDimArray()
     {}
 
+    /// <summary>
+    /// Construct MultiDimArray with dims and array
+    /// </summary>
+    /// <param name="Dims">The dimensions of the array</param>
+    /// <param name="Array_">The array data in fortran order</param>
     public MultiDimArray(uint[] Dims, Array Array_)
     {
 
@@ -36,9 +52,22 @@ public class MultiDimArray
         this.Array_ = Array_;
     }
 
+    /// <summary>
+    /// The dimensions of the array
+    /// </summary>
     public uint[] Dims;
+    /// <summary>
+    /// The data of the array in flattened "Fortran" order
+    /// </summary>
     public Array Array_;
 
+    /// <summary>
+    /// Retrieve a subset of an array
+    /// </summary>
+    /// <param name="memorypos">Position in array to read</param>
+    /// <param name="buffer">Buffer to store retrieved data</param>
+    /// <param name="bufferpos">Position within buffer to store data</param>
+    /// <param name="count">Count of data to retrieve</param>
     public virtual void RetrieveSubArray(uint[] memorypos, MultiDimArray buffer, uint[] bufferpos, uint[] count)
     {
 
@@ -59,6 +88,13 @@ public class MultiDimArray
         }
     }
 
+    /// <summary>
+    /// Assign a subset of an array
+    /// </summary>
+    /// <param name="memorypos">Position within array to store data</param>
+    /// <param name="buffer">Buffer to assign data from</param>
+    /// <param name="bufferpos">Position within buffer to assign from</param>
+    /// <param name="count">Count of data to assign</param>
     public virtual void AssignSubArray(uint[] memorypos, MultiDimArray buffer, uint[] bufferpos, uint[] count)
     {
 
@@ -79,23 +115,55 @@ public class MultiDimArray
         }
     }
 }
-
+/// <summary>
+/// `pod` multidimensional array value type
+/// </summary>
+/// <remarks>
+/// This class stores a pod multidimensional array.
+/// Multidimensional arrays are stored as a uint32_t array of
+/// dimensions, and an array of the flattened elements.
+/// Arrays are stored in column major, or "Fortran" order.
+/// 
+/// Stored type must be a od type that has been generated as part
+/// of the thunk source.
+/// </remarks>
 public class PodMultiDimArray
 {
+    /// <summary>
+    /// Construct empty PodMultiDimArray
+    /// </summary>
     public PodMultiDimArray()
     {
         Dims = new uint[] { 0 };
     }
 
+    /// <summary>
+    /// Construct PodMultiDimArray with dims and array
+    /// </summary>
+    /// <param name="dims">The dimensions of the array</param>
+    /// <param name="array">The array data in fortran order</param>
     public PodMultiDimArray(uint[] dims, Array array)
     {
         Dims = dims;
         pod_array = array;
     }
 
+    /// <summary>
+    /// The dimensions of the array
+    /// </summary>
     public uint[] Dims;
+    /// <summary>
+    /// The data of the array in flattened "Fortran" order
+    /// </summary>
     public Array pod_array;
 
+    /// <summary>
+    /// Retrieve a subset of an array
+    /// </summary>
+    /// <param name="memorypos">Position in array to read</param>
+    /// <param name="buffer">Buffer to store retrieved data</param>
+    /// <param name="bufferpos">Position within buffer to store data</param>
+    /// <param name="count">Count of data to retrieve</param>
     public virtual void RetrieveSubArray(uint[] memorypos, PodMultiDimArray buffer, uint[] bufferpos, uint[] count)
     {
 
@@ -116,6 +184,13 @@ public class PodMultiDimArray
         }
     }
 
+    /// <summary>
+    /// Assign a subset of an array
+    /// </summary>
+    /// <param name="memorypos">Position within array to store data</param>
+    /// <param name="buffer">Buffer to assign data from</param>
+    /// <param name="bufferpos">Position within buffer to assign from</param>
+    /// <param name="count">Count of data to assign</param>
     public virtual void AssignSubArray(uint[] memorypos, PodMultiDimArray buffer, uint[] bufferpos, uint[] count)
     {
         PodMultiDimArray mema = this;
@@ -136,22 +211,51 @@ public class PodMultiDimArray
     }
 }
 
+/// <summary>
+/// `namedarray` multidimensional array value type
+/// </summary>
+/// <remarks>
+/// This class stores a namedarray multidimensional array.
+/// Multidimensional arrays are stored as a uint32_t array of
+/// dimensions, and an array of the flattened elements.
+/// Arrays are stored in column major, or "Fortran" order.
+/// </remarks>
 public class NamedMultiDimArray
 {
+    /// <summary>
+    /// Construct empty NamedMultiDimArray
+    /// </summary>
     public NamedMultiDimArray()
     {
         Dims = new uint[] { 0 };
     }
 
+    /// <summary>
+    /// Construct NamedMultiDimArray with dims and array
+    /// </summary>
+    /// <param name="dims">The dimensions of the array</param>
+    /// <param name="array">The array data in fortran order</param>
     public NamedMultiDimArray(uint[] dims, Array array)
     {
         Dims = dims;
         namedarray_array = array;
     }
 
+    /// <summary>
+    /// The dimensions of the array
+    /// </summary>
     public uint[] Dims;
+    /// <summary>
+    /// The data of the array in flattened "Fortran" order
+    /// </summary>
     public Array namedarray_array;
-
+    /// <summary>
+    /// Retrieve a subset of an array
+    /// </summary>
+    /// <param name="memorypos">Position in array to read</param>
+    /// <param name="buffer">Buffer to store retrieved data</param>
+    /// <param name="bufferpos">Position within buffer to store data</param>
+    /// <param name="count">Count of data to retrieve</param>
     public virtual void RetrieveSubArray(uint[] memorypos, NamedMultiDimArray buffer, uint[] bufferpos, uint[] count)
     {
 
@@ -172,6 +276,13 @@ public class NamedMultiDimArray
         }
     }
 
+    /// <summary>
+    /// Assign a subset of an array
+    /// </summary>
+    /// <param name="memorypos">Position within array to store data</param>
+    /// <param name="buffer">Buffer to assign data from</param>
+    /// <param name="bufferpos">Position within buffer to assign from</param>
+    /// <param name="count">Count of data to assign</param>
     public virtual void AssignSubArray(uint[] memorypos, NamedMultiDimArray buffer, uint[] bufferpos, uint[] count)
     {
         NamedMultiDimArray mema = this;
