+++ RobotRaconteurNET/RobotRaconteurNET/Service.cs
@@ -20,1437 +20,55 @@ using System.Threading.Tasks;
 namespace RobotRaconteur
 {
 
-public abstract class ServiceFactory
-{
-    private ServiceDefinition sdef = null;
-
-    public ServiceDefinition ServiceDef()
-    {
-        if (sdef == null)
-        {
-            sdef = new ServiceDefinition();
-            sdef.FromString(DefString());
-        }
-        return sdef;
-    }
-
-    public string RemovePath(string path)
-    {
-        return RobotRaconteurNode.SplitQualifiedName(path).Item2;
-    }
-
-    public bool CompareNamespace(string qualified_typename, out string typename)
-    {
-        var s = RobotRaconteurNode.SplitQualifiedName(qualified_typename);
-        typename = s.Item2;
-        return s.Item1 == GetServiceName();
-    }
-
-    public abstract string DefString();
-
-    public MessageElementNestedElementList PackStructure(object s)
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.GetType()), out typename))
-        {
-            return FindStructureStub(typename).PackStructure(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackStructure(s);
-        }
-    }
-
-    public T UnpackStructure<T>(MessageElementNestedElementList l)
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return (T)FindStructureStub(typename).UnpackStructure<T>(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackStructure<T>(l);
-        }
-    }
-
-    public MessageElementNestedElementList PackPodToArray<T>(ref T s)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.GetType()), out typename))
-        {
-            return ((PodStub<T>)FindPodStub(typename)).PackPodToArray(ref s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackPodToArray(ref s);
-        }
-    }
-
-    public T UnpackPodFromArray<T>(MessageElementNestedElementList l)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return ((PodStub<T>)FindPodStub(typename)).UnpackPodFromArray(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackPodFromArray<T>(l);
-        }
-    }
-
-    public MessageElementNestedElementList PackPodArray<T>(T[] s)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.GetType().GetElementType()), out typename))
-        {
-            return ((PodStub<T>)FindPodStub(typename)).PackPodArray(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackPodArray(s);
-        }
-    }
-
-    public T[] UnpackPodArray<T>(MessageElementNestedElementList l)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return ((PodStub<T>)FindPodStub(typename)).UnpackPodArray(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackPodArray<T>(l);
-        }
-    }
-
-    public MessageElementNestedElementList PackPodMultiDimArray<T>(PodMultiDimArray s)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.pod_array.GetType().GetElementType()), out typename))
-        {
-            return ((PodStub<T>)FindPodStub(typename)).PackPodMultiDimArray(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackPodMultiDimArray<T>(s);
-        }
-    }
-
-    public PodMultiDimArray UnpackPodMultiDimArray<T>(MessageElementNestedElementList l)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return ((PodStub<T>)FindPodStub(typename)).UnpackPodMultiDimArray(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackPodMultiDimArray<T>(l);
-        }
-    }
-
-    public MessageElementData PackPod(object s)
-    {
-        Type t;
-
-        var s1 = s as PodMultiDimArray;
-        if (s1 != null)
-        {
-            t = s1.pod_array.GetType().GetElementType();
-        }
-        else
-        {
-            var s2 = s as Array;
-            if (s2 != null)
-            {
-                t = s2.GetType().GetElementType();
-            }
-            else
-            {
-                t = s.GetType();
-            }
-        }
-
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(t), out typename))
-        {
-            return FindPodStub(typename).PackPod(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackPod(s);
-        }
-    }
-
-    public object UnpackPod(MessageElementData m)
-    {
-        string typename;
-        if (CompareNamespace(m.GetTypeString(), out typename))
-        {
-            return FindPodStub(typename).UnpackPod(m);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackPod(m);
-        }
-    }
-
-    // namedarray
-
-    public MessageElementNestedElementList PackNamedArrayToArray<T>(ref T s)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.GetType()), out typename))
-        {
-            return ((INamedArrayStub<T>)FindNamedArrayStub(typename)).PackNamedArrayStructToArray(ref s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackNamedArrayToArray(ref s);
-        }
-    }
-
-    public T UnpackNamedArrayFromArray<T>(MessageElementNestedElementList l)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return ((INamedArrayStub<T>)FindNamedArrayStub(typename)).UnpackNamedArrayStructFromArray(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackNamedArrayFromArray<T>(l);
-        }
-    }
-
-    public MessageElementNestedElementList PackNamedArray<T>(T[] s)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.GetType().GetElementType()), out typename))
-        {
-            return ((INamedArrayStub<T>)FindNamedArrayStub(typename)).PackNamedArray(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackNamedArray(s);
-        }
-    }
-
-    public T[] UnpackNamedArray<T>(MessageElementNestedElementList l)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return ((INamedArrayStub<T>)FindNamedArrayStub(typename)).UnpackNamedArray(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackNamedArray<T>(l);
-        }
-    }
-
-    public MessageElementNestedElementList PackNamedMultiDimArray<T>(NamedMultiDimArray s)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(s.namedarray_array.GetType().GetElementType()),
-                             out typename))
-        {
-            return ((INamedArrayStub<T>)FindNamedArrayStub(typename)).PackNamedMultiDimArray(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackNamedMultiDimArray<T>(s);
-        }
-    }
-
-    public NamedMultiDimArray UnpackNamedMultiDimArray<T>(MessageElementNestedElementList l)
-        where T : struct
-    {
-        string typename;
-        if (CompareNamespace(l.GetTypeString(), out typename))
-        {
-            return ((INamedArrayStub<T>)FindNamedArrayStub(typename)).UnpackNamedMultiDimArray(l);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackNamedMultiDimArray<T>(l);
-        }
-    }
-
-    public MessageElementData PackNamedArray(object s)
-    {
-        Type t;
-
-        var s1 = s as NamedMultiDimArray;
-        if (s1 != null)
-        {
-            t = s1.namedarray_array.GetType().GetElementType();
-        }
-        else
-        {
-            var s2 = s as Array;
-            if (s2 != null)
-            {
-                t = s2.GetType().GetElementType();
-            }
-            else
-            {
-                t = s.GetType();
-            }
-        }
-
-        string typename;
-        if (CompareNamespace(RobotRaconteurNode.GetTypeString(t), out typename))
-        {
-            return FindNamedArrayStub(typename).PackNamedArray(s);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.PackNamedArray(s);
-        }
-    }
-
-    public object UnpackNamedArray(MessageElementData m)
-    {
-        string typename;
-        if (CompareNamespace(m.GetTypeString(), out typename))
-        {
-            return FindNamedArrayStub(typename).UnpackNamedArray(m);
-        }
-        else
-        {
-            return RobotRaconteurNode.s.UnpackNamedArray(m);
-        }
-    }
-
-    public abstract IStructureStub FindStructureStub(string objecttype);
-
-    public abstract IPodStub FindPodStub(string objecttype);
-
-    public abstract INamedArrayStub FindNamedArrayStub(string objecttype);
-
-    public abstract ServiceStub CreateStub(WrappedServiceStub innerstub);
-
-    public abstract ServiceSkel CreateSkel(object innerskel);
-
-    public abstract string GetServiceName();
-
-    public abstract RobotRaconteurException DownCastException(RobotRaconteurException exp);
-}
-
-public interface IStructureStub
-{
-    MessageElementNestedElementList PackStructure(Object s);
-
-    T UnpackStructure<T>(MessageElementNestedElementList m);
-}
-
-public interface IPodStub
-{
-    MessageElementData PackPod(object s);
-
-    object UnpackPod(MessageElementData m);
-}
-
-public abstract class PodStub<T> : IPodStub
-    where T : struct
-{
-    public abstract MessageElementNestedElementList PackPod(ref T s);
-
-    public abstract T UnpackPod(MessageElementNestedElementList m);
-
-    public virtual MessageElementNestedElementList PackPodToArray(ref T s2)
-    {
-        using (var mm = new vectorptr_messageelement())
-        {
-            MessageElementUtil.AddMessageElementDispose(
-                mm, MessageElementUtil.NewMessageElementDispose(0, PackPod(ref s2)));
-
-            return new MessageElementNestedElementList(DataTypes.pod_array_t, TypeName, mm);
-        }
-    }
-
-    public virtual MessageElementNestedElementList PackPodArray(T[] s2)
-    {
-        if (s2 == null)
-            return null;
-
-        using (var mm = new vectorptr_messageelement())
-        {
-            for (int i = 0; i < s2.Length; i++)
-            {
-                MessageElementUtil.AddMessageElementDispose(
-                    mm, MessageElementUtil.NewMessageElementDispose(i, PackPod(ref s2[i])));
-            }
-            return new MessageElementNestedElementList(DataTypes.pod_array_t, TypeName, mm);
-        }
-    }
-
-    public virtual MessageElementNestedElementList PackPodMultiDimArray(PodMultiDimArray s3)
-    {
-        if (s3 == null)
-            return null;
-        using (vectorptr_messageelement l = new vectorptr_messageelement())
-        {
-            MessageElementUtil.AddMessageElementDispose(l, "dims", s3.Dims);
-            using (var s4 = PackPodArray((T[])s3.pod_array))
-            {
-                MessageElementUtil.AddMessageElementDispose(l, "array", s4);
-                return new MessageElementNestedElementList(DataTypes.pod_multidimarray_t, TypeName, l);
-            }
-        }
-    }
-
-    public virtual T UnpackPodFromArray(MessageElementNestedElementList s2)
-    {
-        if (s2.TypeName != TypeName)
-            throw new DataTypeException("pod type mismatch");
-        using (vectorptr_messageelement cdataElements = s2.Elements)
-        {
-            if (cdataElements.Count != 1)
-                throw new DataTypeException("pod type mismatch");
-
-            var e = cdataElements[0];
-            using (e)
-            {
-                if (0 != MessageElementUtil.GetMessageElementNumber(e))
-                    throw new DataTypeException("Error in list format");
-                using (MessageElementNestedElementList md = (MessageElementNestedElementList)e.Data)
-                {
-                    return UnpackPod(md);
-                }
-            }
-        }
-    }
-
-    public virtual T[] UnpackPodArray(MessageElementNestedElementList s2)
-    {
-        if (s2.TypeName != TypeName)
-            throw new DataTypeException("pod type mismatch");
-        int count = 0;
-        using (vectorptr_messageelement cdataElements = s2.Elements)
-        {
-            T[] o = new T[cdataElements.Count];
-            foreach (MessageElement e in cdataElements)
-            {
-                using (e)
-                {
-                    if (count != MessageElementUtil.GetMessageElementNumber(e))
-                        throw new DataTypeException("Error in list format");
-                    using (MessageElementNestedElementList md = (MessageElementNestedElementList)e.Data)
-                    {
-                        o[count] = UnpackPod(md);
-                    }
-                    count++;
-                }
-            }
-            return o;
-        }
-    }
-
-    public virtual PodMultiDimArray UnpackPodMultiDimArray(MessageElementNestedElementList s3)
-    {
-        if (s3.TypeName != TypeName)
-            throw new DataTypeException("pod type mismatch");
-        var o = new PodMultiDimArray();
-        using (vectorptr_messageelement marrayElements = s3.Elements)
-        {
-            o.Dims = (MessageElementUtil.FindElementAndCast<uint[]>(marrayElements, "dims"));
-            using (var s2 = (MessageElementUtil.FindElementAndCast<MessageElementNestedElementList>(marrayElements,
-                                                                                                    "array")))
-            {
-                o.pod_array = UnpackPodArray(s2);
-            }
-        }
-        return o;
-    }
-
-    public virtual MessageElementData PackPod(object s)
-    {
-        if (s is T)
-        {
-            T s2 = (T)s;
-            return PackPodToArray(ref s2);
-        }
-
-        var s3 = s as T[];
-        if (s3 != null)
-        {
-            return PackPodArray(s3);
-        }
-
-        var s4 = s as PodMultiDimArray;
-        if (s4 != null)
-        {
-            return PackPodMultiDimArray(s4);
-        }
-
-        throw new DataTypeException("Unexpected message element type for PackPod");
-    }
-    public virtual object UnpackPod(MessageElementData m)
-    {
-        /*var m2 = m as MessageElementNestedElementList;
-        if (m2 != null)
-        {
-
-            return UnpackPod(m2);
-        }*/
-
-        var m3 = m as MessageElementNestedElementList;
-        if (m3 != null)
-        {
-            if (m3.Type == DataTypes.pod_array_t)
-            {
-                return UnpackPodArray(m3);
-            }
-            if (m3.Type == DataTypes.pod_multidimarray_t)
-            {
-                return UnpackPodMultiDimArray(m3);
-            }
-        }
-
-        var m4 = m as MessageElementNestedElementList;
-        if (m4 != null)
-        {}
-
-        throw new DataTypeException("Unexpected message element type for UnpackPod");
-    }
-
-    public abstract string TypeName { get; }
-}
-
-public class NamedArrayElementTypeAndCount : System.Attribute
-{
-    public NamedArrayElementTypeAndCount(Type element_type, int element_array_count)
-    {
-        ElementArrayType = element_type;
-        ElementArrayCount = element_array_count;
-    }
-
-    public Type ElementArrayType { get; }
-    public int ElementArrayCount { get; }
-}
-
-public interface INamedArrayStub
-{
-    MessageElementData PackNamedArray(object s);
-
-    object UnpackNamedArray(MessageElementData m);
-}
-
-public interface INamedArrayStub<T> : INamedArrayStub
-{
-    MessageElementNestedElementList PackNamedArrayStructToArray(ref T s2);
-
-    MessageElementNestedElementList PackNamedArray(T[] s2);
-
-    MessageElementNestedElementList PackNamedMultiDimArray(NamedMultiDimArray s3);
-
-    T UnpackNamedArrayStructFromArray(MessageElementNestedElementList s2);
-
-    T[] UnpackNamedArray(MessageElementNestedElementList s2);
-
-    NamedMultiDimArray UnpackNamedMultiDimArray(MessageElementNestedElementList s3);
-}
-
-public abstract class NamedArrayStub<T, U> : INamedArrayStub<T>
-    where T : struct
-{
-    public abstract U[] GetNumericArrayFromNamedArrayStruct(ref T s);
-
-    public abstract T GetNamedArrayStructFromNumericArray(U[] m);
-
-    public abstract U[] GetNumericArrayFromNamedArray(T[] s);
-
-    public abstract T[] GetNamedArrayFromNumericArray(U[] m);
-
-    public virtual MessageElementNestedElementList PackNamedArrayStructToArray(ref T s2)
-    {
-        using (var mm = new vectorptr_messageelement())
-        {
-            MessageElementUtil.AddMessageElementDispose(
-                mm, MessageElementUtil.NewMessageElementDispose("array", GetNumericArrayFromNamedArrayStruct(ref s2)));
-
-            return new MessageElementNestedElementList(DataTypes.namedarray_array_t, TypeName, mm);
-        }
-    }
-
-    public virtual MessageElementNestedElementList PackNamedArray(T[] s2)
-    {
-        if (s2 == null)
-            return null;
-
-        using (var mm = new vectorptr_messageelement())
-        {
-
-            MessageElementUtil.AddMessageElementDispose(
-                mm, MessageElementUtil.NewMessageElementDispose("array", GetNumericArrayFromNamedArray(s2)));
-
-            return new MessageElementNestedElementList(DataTypes.namedarray_array_t, TypeName, mm);
-        }
-    }
-
-    public virtual MessageElementNestedElementList PackNamedMultiDimArray(NamedMultiDimArray s3)
-    {
-        if (s3 == null)
-            return null;
-        using (vectorptr_messageelement l = new vectorptr_messageelement())
-        {
-            MessageElementUtil.AddMessageElementDispose(l, "dims", s3.Dims);
-            MessageElementUtil.AddMessageElementDispose(l, "array", PackNamedArray((T[])s3.namedarray_array));
-            return new MessageElementNestedElementList(DataTypes.namedarray_multidimarray_t, TypeName, l);
-        }
-    }
-
-    public virtual T UnpackNamedArrayStructFromArray(MessageElementNestedElementList s2)
-    {
-        if (s2.TypeName != TypeName)
-            throw new DataTypeException("namedarray type mismatch");
-        using (vectorptr_messageelement cdataElements = s2.Elements)
-        {
-            if (cdataElements.Count != 1)
-                throw new DataTypeException("pod type mismatch");
-
-            var a = MessageElementUtil.FindElementAndCast<U[]>(cdataElements, "array");
-
-            return GetNamedArrayStructFromNumericArray(a);
-        }
-    }
-
-    public virtual T[] UnpackNamedArray(MessageElementNestedElementList s2)
-    {
-        if (s2.TypeName != TypeName)
-            throw new DataTypeException("namedarray type mismatch");
-        using (vectorptr_messageelement cdataElements = s2.Elements)
-        {
-            if (cdataElements.Count != 1)
-                throw new DataTypeException("pod type mismatch");
-
-            var a = MessageElementUtil.FindElementAndCast<U[]>(cdataElements, "array");
-
-            return GetNamedArrayFromNumericArray(a);
-        }
-    }
-
-    public virtual NamedMultiDimArray UnpackNamedMultiDimArray(MessageElementNestedElementList s3)
-    {
-        if (s3.TypeName != TypeName)
-            throw new DataTypeException("pod type mismatch");
-        var o = new NamedMultiDimArray();
-        using (vectorptr_messageelement marrayElements = s3.Elements)
-        {
-            o.Dims = (MessageElementUtil.FindElementAndCast<uint[]>(marrayElements, "dims"));
-            using (var s2 = (MessageElementUtil.FindElementAndCast<MessageElementNestedElementList>(marrayElements,
-                                                                                                    "array")))
-            {
-                o.namedarray_array = UnpackNamedArray(s2);
-            }
-        }
-        return o;
-    }
-
-    public virtual MessageElementData PackNamedArray(object s)
-    {
-        if (s is T)
-        {
-            T s2 = (T)s;
-            return PackNamedArrayStructToArray(ref s2);
-        }
-
-        var s3 = s as T[];
-        if (s3 != null)
-        {
-            return PackNamedArray(s3);
-        }
-
-        var s4 = s as NamedMultiDimArray;
-        if (s4 != null)
-        {
-            return PackNamedMultiDimArray(s4);
-        }
-
-        throw new DataTypeException("Unexpected message element type for PackNamedArray");
-    }
-    public virtual object UnpackNamedArray(MessageElementData m)
-    {
-        /*var m2 = m as MessageElementNestedElementList;
-        if (m2 != null)
-        {
-
-            return UnpackPod(m2);
-        }*/
-
-        var m3 = m as MessageElementNestedElementList;
-        if (m3 != null)
-        {
-            if (m3.Type == DataTypes.namedarray_array_t)
-            {
-                return UnpackNamedArray(m3);
-            }
-            if (m3.Type == DataTypes.namedarray_multidimarray_t)
-            {
-                return UnpackNamedMultiDimArray(m3);
-            }
-        }
-        throw new DataTypeException("Unexpected message element type for UnpackNamedArray");
-    }
-
-    public abstract string TypeName { get; }
-}
-
-public abstract class ServiceStub
-{
-
-    protected async Task<MessageElement> rr_async_PropertyGet(string name, int timeout)
-    {
-        AsyncRequestDirectorImpl d = new AsyncRequestDirectorImpl();
-        int id = RRObjectHeap.AddObject(d);
-
-        rr_innerstub.async_PropertyGet(name, timeout, d, id);
-        return await d.Task;
-    }
-
-    protected async Task rr_async_PropertySet(string name, MessageElement value, int timeout)
-    {
-        AsyncRequestDirectorImpl d = new AsyncRequestDirectorImpl();
-        int id = RRObjectHeap.AddObject(d);
-
-        rr_innerstub.async_PropertySet(name, value, timeout, d, id);
-        await d.Task;
-    }
-
-    protected async Task<MessageElement> rr_async_FunctionCall(string name, vectorptr_messageelement p, int timeout)
-    {
-        AsyncRequestDirectorImpl d = new AsyncRequestDirectorImpl();
-        int id = RRObjectHeap.AddObject(d);
-
-        rr_innerstub.async_FunctionCall(name, p, timeout, d, id);
-        return await d.Task;
-    }
-
-    protected async Task<WrappedGeneratorClient> rr_async_GeneratorFunctionCall(string name, vectorptr_messageelement p,
-                                                                                int timeout)
-    {
-        AsyncGeneratorClientReturnDirectorImpl d = new AsyncGeneratorClientReturnDirectorImpl();
-        int id = RRObjectHeap.AddObject(d);
-
-        rr_innerstub.async_GeneratorFunctionCall(name, p, timeout, d, id);
-        return await d.Task;
-    }
-
-    public WrappedServiceStub rr_innerstub;
-    private WrappedServiceStubDirectorNET director;
-
-    public ServiceStub(WrappedServiceStub innerstub)
-    {
-        director = new WrappedServiceStubDirectorNET(this);
-
-        rr_innerstub = innerstub;
-        int id = RRObjectHeap.AddObject(this);
-        innerstub.SetRRDirector(director, id);
-    }
-
-    ~ServiceStub()
-    {
-        // rr_innerstub.RRClose();
-    }
-
-    public abstract void DispatchEvent(string membername, vectorptr_messageelement m);
-
-    object objref_lock = new object();
-
-    public object FindObjRef(string n)
-    {
-
-        WrappedServiceStub s = rr_innerstub.FindObjRef(n);
-        lock (objref_lock)
-        {
-            if (s.GetObjectHeapID() != 0)
-            {
-                return RRObjectHeap.GetObject(s.GetObjectHeapID());
-            }
-            ServiceFactory f = RobotRaconteurNode.s.GetServiceType(s.RR_objecttype.GetServiceDefinition().Name);
-            return f.CreateStub(s);
-        }
-    }
-
-    public object FindObjRef(string n, string i)
-    {
-
-        WrappedServiceStub s = rr_innerstub.FindObjRef(n, i);
-        lock (objref_lock)
-        {
-            if (s.GetObjectHeapID() != 0)
-            {
-                return RRObjectHeap.GetObject(s.GetObjectHeapID());
-            }
-            ServiceFactory f = RobotRaconteurNode.s.GetServiceType(s.RR_objecttype.GetServiceDefinition().Name);
-            return f.CreateStub(s);
-        }
-    }
-
-    public object FindObjRefTyped(string n, string objecttype)
-    {
-        WrappedServiceStub s = rr_innerstub.FindObjRefTyped(n, objecttype);
-        lock (objref_lock)
-        {
-            if (s.GetObjectHeapID() != 0)
-            {
-                return RRObjectHeap.GetObject(s.GetObjectHeapID());
-            }
-            ServiceFactory f = RobotRaconteurNode.s.GetServiceType(s.RR_objecttype.GetServiceDefinition().Name);
-            return f.CreateStub(s);
-        }
-    }
-
-    public object FindObjRefTyped(string n, string i, string objecttype)
-    {
-        WrappedServiceStub s = rr_innerstub.FindObjRefTyped(n, i, objecttype);
-        lock (objref_lock)
-        {
-            if (s.GetObjectHeapID() != 0)
-            {
-                return RRObjectHeap.GetObject(s.GetObjectHeapID());
-            }
-            ServiceFactory f = RobotRaconteurNode.s.GetServiceType(s.RR_objecttype.GetServiceDefinition().Name);
-            return f.CreateStub(s);
-        }
-    }
-
-    public async Task<object> AsyncFindObjRef(string n, int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
-    {
-        AsyncStubReturnDirectorImpl<object> d = new AsyncStubReturnDirectorImpl<object>(null);
-        int id = RRObjectHeap.AddObject(d);
-        rr_innerstub.async_FindObjRef(n, timeout, d, id);
-        return await d.Task;
-    }
-
-    public async Task<object> AsyncFindObjRef(string n, string i, int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
-    {
-        AsyncStubReturnDirectorImpl<object> d = new AsyncStubReturnDirectorImpl<object>(null);
-        int id = RRObjectHeap.AddObject(d);
-        rr_innerstub.async_FindObjRef(n, i, timeout, d, id);
-        return await d.Task;
-    }
-
-    public async Task<T> AsyncFindObjRefTyped<T>(string n, string type,
-                                                 int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
-    {
-        AsyncStubReturnDirectorImpl<T> d = new AsyncStubReturnDirectorImpl<T>(null);
-        int id = RRObjectHeap.AddObject(d);
-        rr_innerstub.async_FindObjRefTyped(n, type, timeout, d, id);
-        return await d.Task;
-    }
-
-    public async Task<T> AsyncFindObjRefTyped<T>(string n, string i, string type,
-                                                 int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
-    {
-        AsyncStubReturnDirectorImpl<T> d = new AsyncStubReturnDirectorImpl<T>(null);
-        int id = RRObjectHeap.AddObject(d);
-        rr_innerstub.async_FindObjRefTyped(n, i, type, timeout, d, id);
-        return await d.Task;
-    }
-
-    public virtual MessageElement CallbackCall(string membername, vectorptr_messageelement m)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-}
-
-public class WrappedServiceStubDirectorNET : WrappedServiceStubDirector
-{
-    WeakReference stub;
-
-    public WrappedServiceStubDirectorNET(ServiceStub stub)
-    {
-        this.stub = new WeakReference(stub);
-    }
-
-    public override MessageElement CallbackCall(string CallbackName, vectorptr_messageelement args)
-    {
-
-        try
-        {
-            using (args)
-            {
-                return ((ServiceStub)stub.Target).CallbackCall(CallbackName, args);
-            }
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override void DispatchEvent(string EventName, vectorptr_messageelement args)
-    {
-        try
-        {
-            using (args)
-            {
-                ((ServiceStub)stub.Target).DispatchEvent(EventName, args);
-            }
-        }
-        catch
-        {}
-    }
-}
-
+/// <summary>
+/// Interface for service objects to receive service notifications
+/// </summary>
+/// <remarks>
+/// Service objects are passed to the service, either when the service is registered
+/// or using objrefs. The service initialized the object by configuring events,
+/// pipes, callbacks, and wires for use. The object may implement IRRServiceObject
+/// to receive notification of when this process is complete, and to receive
+/// a ServerContextPtr and the service path of the object.
+/// IRRServiceObject.RRServiceObjectInit() is called after the object has been
+/// initialized to provide this information.
+/// </remarks>
 public interface IRRServiceObject
 {
+    /// <summary>
+    /// Function called after service object has been initialized
+    /// </summary>
+    /// <remarks>
+    /// Override in the service object to receive notification the service object has
+    /// been initialized, a ServerContextPtr, and the service path.
+    /// </remarks>
+    /// <param name="context">The ServerContextPtr owning the object</param>
+    /// <param name="service_path">The object service path</param>
     void RRServiceObjectInit(ServerContext context, string service_path);
 }
 
-public abstract class ServiceSkel : WrappedServiceSkelDirector
-{
 
-    protected WrappedServiceSkel innerskel;
-
-    public ServiceSkel(object o)
-    {
-        uncastobj = o;
-        if (o == null)
-            throw new NullReferenceException();
-    }
-
-    public virtual void InitCallbackServers(object o)
-    {}
-
-    public string ServicePath
-    {
-        get {
-            return "";
-        }
-    }
-
-    protected internal object uncastobj;
-
-    public object UncastObject
-    {
-        get {
-            return uncastobj;
-        }
-    }
-
-    public abstract MessageElement CallGetProperty(string name, WrappedServiceSkelAsyncAdapter async_adapter);
-
-    public abstract void CallSetProperty(string name, MessageElement m, WrappedServiceSkelAsyncAdapter async_adapter);
-
-    public abstract MessageElement CallFunction(string name, vectorptr_messageelement m,
-                                                WrappedServiceSkelAsyncAdapter async_adapter);
-
-    public abstract object GetSubObj(string name, string ind);
-
-    public virtual void RegisterEvents(object obj1)
-    {}
-
-    public virtual void UnregisterEvents(object obj1)
-    {}
-
-    public virtual void InitPipeServers(object obj1)
-    {}
-    public virtual void InitWireServers(object obj1)
-    {}
-
-    public override void ReleaseCastObject()
-    {
-        // RRObjectHeap.DeleteObject(innerskelid);
-        innerskel = null;
-    }
-
-    public virtual object GetCallbackFunction(uint endpoint, string membername)
-    {
-        throw new MemberNotFoundException("Callback " + membername + " not found");
-    }
-
-    // public virtual MessageEntry CallMemoryFunction(MessageEntry m, Endpoint e) { throw new
-    // MemberNotFoundException("Memory " + m.MemberName + " not found"); }
-
-    internal int innerskelid;
-
-    public override void Init(WrappedServiceSkel skel)
-    {
-        innerskel = skel;
-
-        RegisterEvents(uncastobj);
-        InitPipeServers(uncastobj);
-        InitCallbackServers(uncastobj);
-        InitWireServers(uncastobj);
-
-        var init_object = uncastobj as IRRServiceObject;
-        if (init_object != null)
-        {
-            init_object.RRServiceObjectInit(skel.GetContext(), skel.GetServicePath());
-        }
-    }
-    public override MessageElement _CallFunction(string name, vectorptr_messageelement args,
-                                                 WrappedServiceSkelAsyncAdapter async_adapter)
-    {
-        try
-        {
-            using (args)
-            {
-                return CallFunction(name, args, async_adapter);
-            }
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override MessageElement _CallGetProperty(string name, WrappedServiceSkelAsyncAdapter async_adapter)
-    {
-        try
-        {
-            return CallGetProperty(name, async_adapter);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override void _CallSetProperty(string name, MessageElement m, WrappedServiceSkelAsyncAdapter async_adapter)
-    {
-        try
-        {
-            using (m)
-            {
-                CallSetProperty(name, m, async_adapter);
-            }
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-            }
-        }
-    }
-
-    public abstract string RRType { get; }
-
-    public override WrappedRRObject _GetSubObj(string name, string index)
-    {
-        try
-        {
-            object newobj = GetSubObj(name, index);
-            if (newobj == null)
-                return null;
-            ServiceSkel skel = RobotRaconteurNode.s.GetServiceType(RobotRaconteurNode.SplitQualifiedName(RRType).Item1)
-                                   .CreateSkel(newobj);
-
-            int id = RRObjectHeap.AddObject(skel);
-            skel.innerskelid = id;
-            return new WrappedRRObject(skel.RRType, skel, id);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override void MonitorEnter(int timeout)
-    {
-        try
-        {
-
-            if (!(uncastobj is IRobotRaconteurMonitorObject))
-            {
-                throw new InvalidOperationException("Object is not monitor lockable");
-            }
-
-            ((IRobotRaconteurMonitorObject)uncastobj).RobotRaconteurMonitorEnter(timeout);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-            }
-        }
-    }
-
-    public override void MonitorExit()
-    {
-        try
-        {
-            if (!(uncastobj is IRobotRaconteurMonitorObject))
-            {
-                throw new InvalidOperationException("Object is not monitor lockable");
-            }
-
-            ((IRobotRaconteurMonitorObject)uncastobj).RobotRaconteurMonitorExit();
-        }
-
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-            }
-        }
-    }
-
-    public override WrappedArrayMemoryDirector _GetArrayMemory(string name)
-    {
-        try
-        {
-            return GetArrayMemory(name);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override WrappedMultiDimArrayMemoryDirector _GetMultiDimArrayMemory(string name)
-    {
-        try
-        {
-            return GetMultiDimArrayMemory(name);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public virtual WrappedArrayMemoryDirector GetArrayMemory(string name)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-
-    public virtual WrappedMultiDimArrayMemoryDirector GetMultiDimArrayMemory(string name)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-
-    public override WrappedPodArrayMemoryDirector _GetPodArrayMemory(string name)
-    {
-        try
-        {
-            return GetPodArrayMemory(name);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override WrappedPodMultiDimArrayMemoryDirector _GetPodMultiDimArrayMemory(string name)
-    {
-        try
-        {
-            return GetPodMultiDimArrayMemory(name);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public virtual WrappedPodArrayMemoryDirector GetPodArrayMemory(string name)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-
-    public virtual WrappedPodMultiDimArrayMemoryDirector GetPodMultiDimArrayMemory(string name)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-
-    public override WrappedNamedArrayMemoryDirector _GetNamedArrayMemory(string name)
-    {
-        try
-        {
-            return GetNamedArrayMemory(name);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public override WrappedNamedMultiDimArrayMemoryDirector _GetNamedMultiDimArrayMemory(string name)
-    {
-        try
-        {
-            return GetNamedMultiDimArrayMemory(name);
-        }
-        catch (Exception e)
-        {
-            using (MessageEntry merr = new MessageEntry())
-            {
-                RobotRaconteurExceptionUtil.ExceptionToMessageEntry(e, merr);
-                RRDirectorExceptionHelper.SetError(merr, e.ToString());
-                return null;
-            }
-        }
-    }
-
-    public virtual WrappedNamedArrayMemoryDirector GetNamedArrayMemory(string name)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-
-    public virtual WrappedNamedMultiDimArrayMemoryDirector GetNamedMultiDimArrayMemory(string name)
-    {
-        throw new MemberNotFoundException("Member not found");
-    }
-}
-
-public class ServiceSkelUtil
-{
-    public static Type FindParentInterface(Type objtype)
-    {
-
-        List<Type> interfaces = new List<Type>(objtype.GetInterfaces());
-        interfaces.RemoveAll(
-            x => (x.GetCustomAttributes(typeof(RobotRaconteurServiceObjectInterface), true).Length == 0));
-
-        if (interfaces.Count == 0)
-            throw new DataTypeException("Object not a Robot Raconteur type");
-        if (interfaces.Count == 1)
-            return interfaces[0];
-
-        List<Type> parentinterfaces = new List<Type>();
-
-        for (int i = 0; i < interfaces.Count; i++)
-        {
-            bool parent = true;
-            for (int j = 0; j < interfaces.Count; j++)
-            {
-                if (i != j)
-                    if (interfaces[j].GetInterface(interfaces[i].ToString()) != null)
-                        parent = false;
-            }
-
-            if (parent)
-                parentinterfaces.Add(interfaces[i]);
-        }
-
-        if (parentinterfaces.Count != 1)
-            throw new DataTypeException(
-                "Robot Raconteur types can only directly inheret one Robot Raconteur interface type");
-
-        return parentinterfaces[0];
-    }
-}
-
-public partial class WrappedServiceSkelAsyncAdapter
-{
-    public void EndTask<T>(Task<T> task, Func<T, MessageElement> packer)
-    {
-        if (task.IsCanceled)
-        {
-            var exp = new RobotRaconteur.OperationCancelledException("Async operation was cancelled");
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp);
-            End(error_info);
-            return;
-        }
-
-        if (task.IsFaulted)
-        {
-            var exp = task.Exception;
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp.InnerException);
-            End(error_info);
-            return;
-        }
-
-        MessageElement res;
-        try
-        {
-            res = packer(task.Result);
-        }
-        catch (Exception exp)
-        {
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp);
-            End(error_info);
-            return;
-        }
-
-        End(res, new HandlerErrorInfo());
-    }
-
-    public void EndTask(Task task)
-    {
-        if (task.IsCanceled)
-        {
-            var exp = new RobotRaconteur.OperationCancelledException("Async operation was cancelled");
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp);
-            End(error_info);
-            return;
-        }
-
-        if (task.IsFaulted)
-        {
-            var exp = task.Exception;
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp.InnerException);
-            End(error_info);
-            return;
-        }
-
-        End(new HandlerErrorInfo());
-    }
-
-    public void EndTask(Task task, MessageElement res)
-    {
-        if (task.IsCanceled)
-        {
-            var exp = new RobotRaconteur.OperationCancelledException("Async operation was cancelled");
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp);
-            End(error_info);
-            return;
-        }
-
-        if (task.IsFaulted)
-        {
-            var exp = task.Exception;
-            var error_info = RobotRaconteurExceptionUtil.ExceptionToErrorInfo(exp.InnerException);
-            End(error_info);
-            return;
-        }
-
-        End(res, new HandlerErrorInfo());
-    }
-}
-
-public class RRObjectHeap
-{
-
-    class RRObjectHeap_support : RRNativeObjectHeapSupport
-    {
-        public override void DeleteObject_i(int id)
-        {
-            RRObjectHeap.DeleteObject(id);
-        }
-    }
-
-    static RRObjectHeap_support support;
-
-    static RRObjectHeap()
-    {
-        support = new RRObjectHeap_support();
-        RRNativeObjectHeapSupport.Set_Support(support);
-    }
-
-    public static Dictionary<int, object> objects = new Dictionary<int, object>();
-    public static int objectcount = 0;
-
-    public static int AddObject(object o)
-    {
-        lock (objects)
-        {
-            if (objectcount > int.MaxValue - 10)
-                objectcount = 0;
-            objectcount++;
-            while (objects.ContainsKey(objectcount))
-                objectcount++;
-
-            int i = objectcount;
-            objects.Add(i, o);
-            // Console.WriteLine(objects.Count);
-            return i;
-        }
-    }
-
-    public static void DeleteObject(int id)
-    {
-        lock (objects)
-        {
-            try
-            {
-                objects.Remove(id);
-                // Console.WriteLine(objects.Count);
-            }
-            catch
-            {
-
-                // Console.WriteLine(e.ToString());
-            }
-        }
-    }
-
-    public static object GetObject(int id)
-    {
-        try
-        {
-            return objects[id];
-        }
-        catch (KeyNotFoundException)
-        {
-            throw new ApplicationException("Internal stub error");
-        }
-    }
-}
-
-public class RobotRaconteurServiceObjectInterface : Attribute
-{}
-
+/// <summary>
+/// "callback" member type interface
+/// </summary>
+/// <remarks>
+/// <para>
+/// The Callback class implements the `callback` member type. Callbacks are declared in service definition
+/// files using the `callback` keyword within object declarations. They provide functionality similar to the
+/// `function` member, but the direction is reversed, allowing the service to call a function on a specified
+/// client. The desired client is specified using the Robot Raconteur endpoint identifier. Clients must
+/// configure the callback to use using SetFunction().
+/// </para>
+/// <para>
+///  On the client side,
+/// the client specifies a function for the callback using the SetFunction() function.
+/// On the service side, the function GetFunction(const RR_SHARED_PTR<Endpoint>& endpoint) is used to retrieve
+/// the proxy function to call a client callback.
+/// </para>
+/// <para>
+/// This class is instantiated by the node. It should not be instantiated by the user.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The type of the callback function. This is determined by the thunk source generator.</typeparam>
 public abstract class Callback<T>
 {
 
@@ -1461,8 +79,38 @@ public abstract class Callback<T>
         m_MemberName = name;
     }
 
+    /// <summary>
+    /// Get or set the currently configured callback function on client side
+    /// </summary>
+    /// <remarks>
+    /// The callback function set will be made available to be called by
+    /// the service using a function proxy.
+    /// </remarks>
     public abstract T Function { get; set; }
 
+    /// <summary>
+    /// Get the proxy function to call the callback for the specified client on
+    /// the service side
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    ///  This function returns a proxy to the callback on a specified client. The proxy
+    /// operates as a reverse function, sending parameters, executing the callback,
+    /// and receiving the results.
+    /// </para>
+    /// <para>
+    /// Because services can have multiple clients, it is necessary to specify which client
+    /// to call. This is done by passing the endpoint of the client connection to the
+    /// endpoint parameter.
+    /// </para>
+    /// <para>
+    /// The endpoint of a client can be determined using the ServerEndpoint.CurrentEndpoint()
+    /// function during a `function` or `property` member call. The service can store this
+    /// value, and use it to retrieve the callback proxy.
+    /// </para>
+    /// </remarks>
+    /// <param name="e">The endpoint of the client connection</param>
+    /// <returns>The callback proxy function</returns>
     public abstract T GetClientFunction(uint e);
 }
 
@@ -1513,6 +161,48 @@ public class CallbackServer<T> : Callback<T>
     }
 }
 
+
+/// <summary>
+/// `pipe` member type interface
+/// </summary>
+/// <remarks>
+/// <para>
+/// The Pipe class implements the `pipe` member type. Pipes are declared in service definition files
+/// using the `pipe` keyword within object declarations. Pipes provide reliable packet streaming between
+/// clients and services. They work by creating pipe endpoint pairs (peers), with one endpoint in the client,
+/// and one in the service. Packets are transmitted between endpoint pairs. Packets sent by one endpoint are received
+/// by the other, where they are placed in a receive queue. Received packets can then be retrieved from the receive
+/// queue.
+/// </para>
+/// <para>
+/// Pipe endpoints are created by the client using the Connect() or AsyncConnect() functions. Services receive
+/// incoming connection requests through a callback function. This callback is configured using the
+/// SetPipeConnectCallback() function. Services may also use the PipeBroadcaster class to automate managing pipe endpoint
+/// lifecycles and sending packets to all connected client endpoints. If the SetPipeConnectCallback() function is used,
+/// the service is responsible for keeping track of endpoints as the connect and disconnect. See PipeEndpoint for details
+/// on sending and receiving packets.
+/// </para>
+/// <para>
+/// Pipe endpoints are *indexed*, meaning that more than one endpoint pair can be created between the client and the
+/// service.
+/// </para>
+/// <para>
+///  Pipes may be *unreliable*, meaning that packets may arrive out of order or be dropped. Use IsUnreliable() to check
+/// for unreliable pipes. The member modifier `unreliable` is used to specify that a pipe should be unreliable.
+/// </para>
+/// <para
+/// Pipes may be declared *readonly* or *writeonly*. If neither is specified, the pipe is assumed to be full duplex.
+/// *readonly* pipes may only send packets from service to client. *writeonly* pipes may only send packets from client to
+/// service. Use Direction() to determine the direction of the pipe.
+/// </para>
+/// <para>
+/// The PipeBroadcaster is often used to simplify the use of Pipes. See PipeBroadcaster for more information.
+/// </para>
+/// <para>
+/// This class is instantiated by the node. It should not be instantiated by the user.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The packet data type</typeparam>
 public class Pipe<T>
 {
     internal object innerpipe;
@@ -1522,6 +212,9 @@ public class Pipe<T>
     string m_Membername;
     MemberDefinition_Direction direction;
 
+    /// <summary>
+    /// The pipe member name
+    /// </summary>
     public string MemberName
     {
         get {
@@ -1554,6 +247,14 @@ public class Pipe<T>
         }
     }
 
+    /// <summary>
+    /// Get the direction of the pipe
+    /// </summary>
+    /// <remarks>
+    /// Pipes may be declared *readonly* or* writeonly* in the service definition file. (If neither
+    /// is specified, the pipe is assumed to be full duplex.) *readonly* pipes may only send packets from
+    /// service to client. *writeonly* pipes may only send packets from client to service.
+    /// </remarks>
     public MemberDefinition_Direction Direction
     {
         get {
@@ -1561,6 +262,27 @@ public class Pipe<T>
         }
     }
 
+    /// <summary>
+    /// Connect to any pipe index
+    /// </summary>
+    public const int ANY_INDEX=-1;
+
+    /// <summary>
+    /// Connect a pipe endpoint
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Creates a connected pipe endpoint pair, and returns the local endpoint. Use to create the streaming data
+    /// connection to the service. Pipe endpoints are indexed, meaning that Connect() may be called multiple
+    /// times for the same client connection to create multple pipe endpoint pairs. For most cases Pipe.ANY_INDEX
+    /// (-1) can be used to automatically select an available index.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="index">The index of the pipe endpoint, or (-1) to automatically select an index</param>
+    /// <returns></returns>
     public PipeEndpoint Connect(int index = -1)
     {
 
@@ -1611,11 +333,38 @@ public class Pipe<T>
         }
     }
 
+    /// <summary>
+    /// Asynchronously connect a pipe endpoint.
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Same as Connect(), but returns async.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>Task that upon completion returns connected endpoint</returns>
     public Task<PipeEndpoint> AsyncConnect(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
         return AsyncConnect(-1, timeout);
     }
 
+    /// <summary>
+    /// Asynchronously connect a pipe endpoint.
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Same as Connect(), but returns async.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="index">The index of the pipe endpoint, or ANY_INDEX to automatically select an index</param>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>Task that upon completion returns connected endpoint</returns>
     public async Task<PipeEndpoint> AsyncConnect(int index, int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
 
@@ -1635,6 +384,26 @@ public class Pipe<T>
 
     // public delegate void PipePacketAckReceivedCallbackFunction(PipeEndpoint e, uint packetnum);
 
+    /// <summary>
+    /// Set the pipe endpoint connected callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Callback function invoked when a client attempts to connect a pipe endpoint. The callback
+    /// will receive the incoming pipe endpoint as a parameter. The service must maintain a reference to the
+    /// pipe endpoint, but the pipe will retain ownership of the endpoint until it is closed.
+    /// </para>
+    /// </remarks>
+    /// <para>
+    /// The callback may throw an exception to reject incoming connect request.
+    /// </para>
+    /// <para>
+    /// Note: Connect callback is configured automatically by PipeBroadcaster
+    /// </para>
+    /// <para>
+    /// Only valid for services. Will throw InvalidOperationException on the client side.
+    /// </para>
+    /// <value></value>
     public Action<PipeEndpoint> PipeConnectCallback
     {
         get {
@@ -1675,11 +444,64 @@ public class Pipe<T>
         }
     }
 
+    /// <summary>
+    /// Pipe endpoint used to transmit reliable or unreliable data streams
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Pipe endpoints are used to communicate data between connected pipe members.
+    /// See Pipe for more information on pipe members.
+    /// </para>
+    /// <para>
+    /// Pipe endpoints are created by clients using the Pipe.Connect() or Pipe.AsyncConnect()
+    /// functions. Services receive incoming pipe endpoint connection requests through a
+    /// callback function specified using the Pipe.PipeConnectCallback property. Services
+    /// may also use the PipeBroadcaster class to automate managing pipe endpoint lifecycles and
+    /// sending packets to all connected client endpoints.
+    /// </para>
+    /// <para>
+    /// Pipe endpoints are///indexed*, meaning that more than one pipe endpoint pair can be created
+    /// using the same member. This means that multiple data streams can be created independent of
+    /// each other between the client and service using the same member.
+    /// </para>
+    /// <para>
+    /// Pipes send reliable packet streams between connected client/service endpoint pairs.
+    /// Packets are sent using the SendPacket() or AsyncSendPacket() functions. Packets
+    /// are read from the receive queue using the ReceivePacket(), ReceivePacketWait(),
+    /// TryReceivePacketWait(), TryReceivePacketWait(), or PeekNextPacket(). The endpoint is closed
+    /// using the Close() or AsyncClose() function.
+    /// </para>
+    /// <para>
+    /// This class is instantiated by the Pipe class. It should not be instantiated
+    /// by the user.
+    /// </para>
+    /// </remarks>
     public class PipeEndpoint
     {
         WrappedPipeEndpoint innerpipe;
 
+        /// <summary>
+        /// Signal called when a packet has been received
+        /// </summary>
+        /// <remarks>
+        /// Function must accept one argument, receiving the PipeEndpoint that
+        /// received a packet
+        /// </remarks>
         public event Action<PipeEndpoint> PacketReceivedEvent;
+
+        /// <summary>
+        /// Signal called when a packet ack has been received
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Packet acks are generated if SetRequestPacketAck() is set to true. The receiving
+        /// endpoint generates acks to inform the sender that the packet has been received.
+        /// </para>
+        /// <para>
+        /// Function must accept two arguments, receiving the PipeEndpoint
+        /// that received the packet ack and the packet number that is being acked.
+        /// </para>
+        /// </remarks>
         public event Action<PipeEndpoint, uint> PacketAckReceivedEvent;
 
         directorclass director;
@@ -1700,6 +522,9 @@ public class Pipe<T>
             // RRObjectHeap.DeleteObject(director_id);
         }
 
+        /// <summary>
+        /// Get the pipe endpoint index used when endpoint connected
+        /// </summary>
         public int Index
         {
             get {
@@ -1707,6 +532,13 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Get the Robot Raconteur node Endpoint ID
+        /// </summary>
+        /// <remarks>
+        /// Get the endpoint associated with the ClientContext or ServerEndpoint
+        /// associated with the pipe endpoint.
+        /// </remarks>
         public uint Endpoint
         {
             get {
@@ -1714,6 +546,14 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Get or set if pipe endpoint should request packet acks
+        /// </summary>
+        /// <remarks>
+        /// Packet acks are generated by receiving endpoints to inform the sender that
+        /// a packet has been received. The ack contains the packet index, the sequence number
+        /// of the packet. Packet acks are used for flow control by PipeBroadcaster.
+        /// </remarks>
         public bool RequestPacketAck
         {
             get {
@@ -1724,6 +564,15 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Get or set whether pipe endpoint should ignore incoming packets
+        /// </summary>
+        /// <remarks>
+        /// Pipe endpoints may optionally desire to ignore incoming data. This is useful if the endpoint
+        /// is only being used to send packets, and received packets may create a potential memory leak if they
+        /// are not being removed from the queue. If ignore is true, incoming packets will be discarded and
+        /// will not be added to the receive queue.
+        /// </remarks>
         public bool IgnoreReceived
         {
             get {
@@ -1734,6 +583,14 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// The direction of the pipe
+        /// </summary>
+        /// <remarks>
+        /// Pipes may be declared*readonly* or*writeonly* in the service definition file. (If neither
+        /// is specified, the pipe is assumed to be full duplex.)*readonly* pipes may only send packets from
+        /// service to client.*writeonly* pipes may only send packets from client to service.
+        /// </remarks>
         public MemberDefinition_Direction Direction
         {
             get {
@@ -1743,6 +600,18 @@ public class Pipe<T>
 
         private Action<PipeEndpoint> close_callback;
 
+        /// <summary>
+        /// Get or set the endpoint closed callback function
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Get or Set a function to invoke when the pipe endpoint has been closed.
+        /// </para>
+        /// <para>
+        /// Callback function must accept one argument, receiving the PipeEndpoint that
+        /// was closed.
+        /// </para>
+        /// </remarks>        
         public Action<PipeEndpoint> PipeCloseCallback
         {
             get {
@@ -1753,6 +622,12 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Return number of packets in the receive queue
+        /// </summary>
+        /// <remarks>
+        /// Invalid for writeonly pipes.
+        /// </remarks>
         public int Available
         {
             get {
@@ -1760,6 +635,13 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Get if pipe endpoint is unreliable
+        /// </summary>
+        /// <remarks>
+        /// Pipe members may be declared as*unreliable* using member modifiers in the
+        /// service definition. Pipes confirm unreliable operation when pipe endpoints are connected.
+        /// </remarks>
         public bool IsUnreliable
         {
             get {
@@ -1767,6 +649,15 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Peeks the next packet in the receive queue
+        /// </summary>
+        /// <remarks>
+        /// Returns the first packet in the receive queue, but does not remove it from
+        /// the queue. Throws an InvalidOperationException if there are no packets in the
+        /// receive queue.
+        /// </remarks>
+        /// <returns>The next packet in the receive queue</returns>
         public T PeekNextPacket()
         {
             using (MessageElement m = innerpipe.PeekNextPacket())
@@ -1786,6 +677,15 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Receive the next packet in the receive queue
+        /// </summary>
+        /// <remarks>
+        /// Receive the next packet from the receive queue. This function will throw an
+        /// InvalidOperationException if there are no packets in the receive queue. Use
+        /// ReceivePacketWait() to block until a packet has been received.
+        /// </remarks>
+        /// <returns>The received packet</returns>
         public T ReceivePacket()
         {
             using (MessageElement m = innerpipe.ReceivePacket())
@@ -1805,6 +705,14 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Peek the next packet in the receive queue, block if queue is empty
+        /// </summary>
+        /// <remarks>
+        /// Same as PeekPacket(), but blocks if queue is empty
+        /// </remarks>
+        /// <param name="timeout">Timeout in milliseconds to wait for a packet, or RR_TIMEOUT_INFINITE for no timeout</param>
+        /// <returns>The received packet</returns>
         public T PeekNextPacketWait(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
         {
             using (MessageElement m = innerpipe.PeekNextPacketWait(timeout))
@@ -1824,6 +732,14 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Receive the next packet in the receive queue, block if queue is empty
+        /// </summary>
+        /// <remarks>
+        /// Same as ReceivePacket(), but blocks if queue is empty
+        /// </remarks>
+        /// <param name="timeout">Timeout in milliseconds to wait for a packet, or RR_TIMEOUT_INFINITE for no timeout</param>
+        /// <returns>The received packet</returns>
         public T ReceivePacketWait(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
         {
             using (MessageElement m = innerpipe.ReceivePacketWait(timeout))
@@ -1843,6 +759,22 @@ public class Pipe<T>
             }
         }
 
+        /// <summary>
+        /// Try receiving a packet, optionally blocking if the queue is empty
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Try receiving a packet with various options. Returns true if a packet has been
+        /// received, or false if no packet is available instead of throwing an exception on failure.
+        /// The timeout and peek parameters can be used to modify behavior to provide functionality
+        /// similar to the various Receive and Peek functions.
+        /// </para>
+        /// </remarks>
+        /// <param name="packet">[out] The received packet</param>
+        /// <param name="timeout">The timeout in milliseconds. Set to zero for non-blocking operation, an arbitrary value
+        ///     in milliseconds for a finite duration timeout, or RR_TIMEOUT_INFINITE for no timeout</param>
+        /// <param name="peek">If true, the packet is not removed from the receive queue</param>
+        /// <returns>true if packet was received, otherwise false</returns>
         public bool TryReceivePacketWait(out T packet, int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE,
                                          bool peek = false)
         {
@@ -1869,12 +801,23 @@ public class Pipe<T>
             }
         }
 
-        public uint SendPacket(T data)
+        /// <summary>
+        /// Sends a packet to the peer endpoint
+        /// </summary>
+        /// <remarks>
+        /// Sends a packet to the peer endpoint. If the pipe is reliable, the packetsare  guaranteed to arrive
+        /// in order. If the pipe is set to unreliable, "best effort" is made to deliver packets, and they are not
+        /// guaranteed to arrive in order. This function will block until the packet has been transmitted by the
+        /// transport. It will return before the peer endpoint has received the packet.
+        /// </remarks>
+        /// <param name="packet">The packet to send</param>
+        /// <returns></returns>
+        public uint SendPacket(T packet)
         {
 
             try
             {
-                using (MessageElement m = RobotRaconteurNode.s.PackAnyType<T>("value", ref data))
+                using (MessageElement m = RobotRaconteurNode.s.PackAnyType<T>("value", ref packet))
                 {
                     return innerpipe.SendPacket(m);
                 }
@@ -1883,14 +826,22 @@ public class Pipe<T>
             {}
         }
 
-        public async Task<uint> AsyncSendPacket(T data)
+        /// <summary>
+        /// Send a packet to the peer endpoint asynchronously
+        /// </summary>
+        /// <remarks>
+        /// Same as SendPacket(), but async.
+        /// </remarks>
+        /// <param name="packet">The packet to send</param>
+        /// <returns>A task that upon completion returns the sent packet number</returns>
+        public async Task<uint> AsyncSendPacket(T packet)
         {
 
             AsyncUInt32ReturnDirectorImpl h = new AsyncUInt32ReturnDirectorImpl();
 
             try
             {
-                using (MessageElement m = RobotRaconteurNode.s.PackAnyType<T>("value", ref data))
+                using (MessageElement m = RobotRaconteurNode.s.PackAnyType<T>("value", ref packet))
                 {
                     int id = RRObjectHeap.AddObject(h);
                     innerpipe.AsyncSendPacket(m, h, id);
@@ -1901,11 +852,25 @@ public class Pipe<T>
             return await h.Task;
         }
 
+        /// <summary>
+        /// Close the pipe endpoint
+        /// </summary>
+        /// <remarks>
+        ///  Close the pipe endpoint. Blocks until close complete. The peer endpoint is destroyed
+        /// automatically.
+        /// </remarks>
         public void Close()
         {
             innerpipe.Close();
         }
 
+        /// <summary>
+        /// Asynchronously close the pipe endpoint
+        /// </summary>
+        /// <remarks>
+        /// Same as Close() but async
+        /// </remarks>
+        /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
         public async Task AsyncClose(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
         {
             AsyncVoidReturnDirectorImpl h = new AsyncVoidReturnDirectorImpl();
@@ -1980,6 +945,43 @@ public class Pipe<T>
     }
 }
 
+/// <summary>
+/// Broadcaster to send packets to all connected clients
+/// </summary>
+/// <remarks>
+/// <para>
+/// PipeBroadcaster is used by services to send packets to all connected
+/// client endpoints. It attaches to the pipe on the service side, and
+/// manages the lifecycle of connected endpoints. PipeBroadcaster should
+/// only be used with pipes that are declared*readonly*, since it has
+/// no provisions for receiving incoming packets from the client.
+/// </para>
+/// <para>
+/// PipeBroadcaster is initialized by the user, or by default implementation
+/// classes generated by RobotRaconteurGen (*_default_impl). Default
+/// implementation classes will automatically instantiate broadcasters
+/// for pipes marked*readonly*. If default implementation classes are
+/// not used, the broadcaster must be instantiated manually. It is
+/// recommended this be done using the IRRServiceObject interface in
+/// the overridden IRRServiceObject.RRServiceObjectInit() function. This
+/// function is called after the pipes have been instantiated by the service.
+/// </para>
+/// <para>
+/// Use SendPacket() or AsyncSendPacket() to broadcast packets to all
+/// connected clients.
+/// </para>
+/// <para>
+/// PipeBroadcaster provides flow control by optionally tracking how many packets
+/// are in flight to each client pipe endpoint. (This is accomplished using packet acks.) If a
+/// maximum backlog is specified, pipe endpoints exceeding this count will stop sending packets.
+/// Specify the maximum backlog using the Init() function or the SetMaxBacklog() function.
+/// </para>
+/// <para>
+/// The rate that packets are sent can be regulated using a callback function configured
+/// with the Predicate property, or using the BroadcastDownsampler class.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The packet data type</typeparam>
 public class PipeBroadcaster<T>
 {
     protected Pipe<T> pipe;
@@ -2001,6 +1003,12 @@ public class PipeBroadcaster<T>
         }
     }
 
+    /// <summary>
+    /// Construct a new PipeBroadcaster
+    /// </summary>
+    /// <param name="pipe">The pipe to use for broadcasting. Must be a pipe from a service object.
+    /// Specifying a client pipe will result in an exception.</param>
+    /// <param name="maximum_backlog">The maximum number of packets in flight, or -1 for unlimited</param>
     public PipeBroadcaster(Pipe<T> pipe, int maximum_backlog = -1)
     {
         this.pipe = pipe;
@@ -2009,6 +1017,9 @@ public class PipeBroadcaster<T>
         this.type = ((WrappedPipeServer)pipe.innerpipe).Type;
     }
 
+    /// <summary>
+    /// Get the associated pipe
+    /// </summary>
     public Pipe<T> Pipe
     {
         get {
@@ -2016,6 +1027,13 @@ public class PipeBroadcaster<T>
         }
     }
 
+    /// <summary>
+    /// Asynchronously send packet to all connected pipe endpoint clients
+    /// </summary>
+    /// <remarks>
+    /// Async version of SendPacket()
+    /// </remarks>
+    /// <param name="packet">The packet to send</param>
     public async Task AsyncSendPacket(T packet)
     {
         AsyncVoidNoErrReturnDirectorImpl h = new AsyncVoidNoErrReturnDirectorImpl();
@@ -2034,6 +1052,13 @@ public class PipeBroadcaster<T>
         await h.Task;
     }
 
+    /// <summary>
+    /// Send a packet to all connected pipe endpoint clients
+    /// </summary>
+    /// <remarks>
+    /// Blocks until packet has been sent by all endpoints
+    /// </remarks>
+    /// <param name="packet">The packet to send</param>
     public void SendPacket(T packet)
     {
         try
@@ -2048,6 +1073,9 @@ public class PipeBroadcaster<T>
         {}
     }
 
+    /// <summary>
+    /// Get the number of active client endpoints
+    /// </summary>
     public uint ActivePipeEndpointCount
     {
         get {
@@ -2055,6 +1083,25 @@ public class PipeBroadcaster<T>
         }
     }
 
+
+    /// <summary>
+    /// Set the predicate callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// A predicate is optionally used to regulate when packets are sent to clients. This is used by the
+    /// BroadcastDownsampler to regulate update rates of packets sent to clients.
+    /// </para>
+    /// <para>
+    /// The predicate callback is invoked before the broadcaster sends a packet to an endpoint. If the predicate returns
+    /// true, the packet will be sent. If it is false, the packet will not be sent to that endpoint.
+    /// </para>
+    /// <para>
+    /// It receives the broadcaster, the client endpoint ID, and the pipe endpoint index. It returns true to send the
+    /// packet, or false to not send the packet.
+    /// </para>
+    /// </remarks>
+    /// <value></value>
     public Func<uint, int, bool> Predicate
     {
         set {
@@ -2064,6 +1111,16 @@ public class PipeBroadcaster<T>
         }
     }
 
+
+    /// <summary>
+    /// Get or set the maximum backlog
+    /// </summary>
+    /// <remarks>
+    /// PipeBroadcaster provides flow control by optionally tracking how many packets
+    /// are in flight to each client pipe endpoint. (This is accomplished using packet acks.) If a
+    /// maximum backlog is specified, pipe endpoints exceeding this count will stop sending packets.
+    /// Use -1 for infinite packets in flight.
+    /// </remarks>
     public int MaxBacklog
     {
         get {
@@ -2075,6 +1132,67 @@ public class PipeBroadcaster<T>
     }
 }
 
+/// <summary>
+/// `wire` member type interface
+/// </summary>
+/// <remarks>
+/// <para>
+/// The Wire class implements the `wire` member type. Wires are declared in service definition files
+/// using the `wire` keyword within object declarations. Wires provide "most recent" value streaming
+/// between clients and services. They work by creating "connection" pairs between the client and service.
+/// The wire streams the current value between the wire connection pairs using packets. Wires
+/// are unreliable; only the most recent value is of interest, and any older values
+/// will be dropped. Wire connections have an InValue and an OutValue. Users set the OutValue on the
+/// connection. The new OutValue is transmitted to the peer wire connection, and becomes the peer's
+/// InValue. The peer can then read the InValue. The client and service have their own InValue
+/// and OutValue, meaning that each direction, client to service or service to client, has its own
+/// value.
+/// </para>
+/// <para>
+/// Wire connections are created using the Connect() or AsyncConnect() functions. Services receive
+/// incoming connection requests through a callback function. Thes callback is configured using
+/// the SetWireConnectCallback() function. Services may also use the WireBroadcaster class
+/// or WireUnicastReceiver class to automate managing wire connection lifecycles. WireBroadcaster
+/// is used to send values to all connected clients. WireUnicastReceiver is used to receive the
+/// value from the most recent wire connection. See WireConnection for details on sending
+/// and receiving streaming values.
+/// </para>
+/// <para>
+/// Wire clients may also optionally "peek" and "poke" the wire without forming a streaming
+/// connection. This is useful if the client needs to read the InValue or set the OutValue
+/// instantaniously, but does not need continuous updating. PeekInValue() or
+/// AsyncPeekInValue() will retrieve the client's current InValue. PokeOutValue() or
+/// AsyncPokeOutValue() will send a new client OutValue to the service.
+/// PeekOutValue() or AsyncPeekOutValue() will retrieve the last client OutValue received by
+/// the service.
+/// </para>
+/// <para>
+/// "Peek" and "poke" operations initiated by the client are received on the service using
+/// callbacks. Use SetPeekInValueCallback(), SetPeekOutValueCallback(),
+/// and SetPokeOutValueCallback() to configure the callbacks to handle these requests.
+/// WireBroadcaster and WireUnicastReceiver configure these callbacks automatically, so
+/// the user does not need to configure the callbacks when these classes are used.
+/// </para>
+/// <para>
+/// Wires can be declared*readonly* or*writeonly*. If neither is specified, the wire is assumed
+/// to be full duplex.*readonly* pipes may only send values from service to client, ie OutValue
+/// on service side and InValue on client side.*writeonly* pipes may only send values from
+/// client to service, ie OutValue on client side and InValue on service side. Use Direction()
+/// to determine the direction of the wire.
+/// </para>
+/// <para>
+/// Unlike pipes, wire connections are not indexed, so only one connection pair can be
+/// created per client connection.
+/// </para>
+/// <para>
+/// WireBroadcaster or WireUnicastReceiver are typically used to simplify using wires.
+/// See WireBroadcaster and WireUnicastReceiver for more information.
+/// </para>
+/// <para>
+/// This class is instantiated by the node. It should not be instantiated by the user.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The value data type</typeparam>
 public class Wire<T>
 {
     internal object innerwire;
@@ -2082,6 +1200,9 @@ public class Wire<T>
     string m_Membername;
     MemberDefinition_Direction direction;
 
+    /// <summary>
+    /// Get the member name of the wire
+    /// </summary>
     public string MemberName
     {
         get {
@@ -2089,6 +1210,14 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Get the direction of the wire
+    /// </summary>
+    /// <remarks>
+    /// Wires may be declared*readonly* or*writeonly* in the service definition file. (If neither
+    /// is specified, the wire is assumed to be full duplex.)*readonly* wire may only send out values from
+    /// service to client.*writeonly* wires may only send out values from client to service.
+    /// </remarks>
     public MemberDefinition_Direction Direction
     {
         get {
@@ -2120,6 +1249,23 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Connect the wire
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Creates a connection between the wire, returning the client connection. Used to create
+    /// a "most recent" value streaming connection to the service.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// <para>
+    /// Note: If a streaming connection is not required, use PeekInValue(), PeekOutValue(),
+    /// or PokeOutValue() instead of creating a connection.
+    /// </para>
+    /// </remarks>
+    /// <returns>The wire connection</returns>
     public WireConnection Connect()
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2170,6 +1316,19 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Asynchronously connect the wire
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Same as Connect(), but async
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>Task that upon completion returns the connected WireConnection</returns>
     public async Task<WireConnection> AsyncConnect(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2181,6 +1340,29 @@ public class Wire<T>
         return await h.Task;
     }
 
+
+    /// <summary>
+    /// Set wire connected callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Callback function invoked when a client attempts to connect a the wire. The callback
+    /// will receive the incoming wire connection as a parameter. The service must maintain a
+    /// reference to the wire connection, but the wire will retain ownership of the wire connection
+    /// until it is closed. Using  boost::weak_ptr to store the reference to the connection
+    /// is recommended.
+    /// </para>
+    /// <para>
+    /// The callback may throw an exception to reject incoming connect request.
+    /// </para>
+    /// <para>
+    /// Note: Connect callback is configured automatically by WireBroadcaster or
+    /// WireUnicastReceiver
+    /// </para>
+    /// <para>
+    /// Only valid for services. Will throw InvalidOperationException on the client side.
+    /// </para>
+    /// </remarks>
     public Action<Wire<T>, WireConnection> WireConnectCallback
     {
         get {
@@ -2274,6 +1456,27 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Peek the current InValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Peeks the current InValue using a "request" instead of a streaming value. Use
+    /// if only the instantanouse value is required.
+    /// </para>
+    /// <para>
+    /// Peek and poke are similar to `property` members. Unlike streaming,
+    /// peek and poke are reliable operations.
+    /// </para>
+    /// <para>
+    /// Throws ValueNotSetException if InValue is not valid.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="ts">[out] The timestamp of the current InValue</param>
+    /// <returns>The current InValue</returns>
     public T PeekInValue(out TimeSpec ts)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2286,6 +1489,27 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Peek the current OutValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Peeks the current OutValue using a "request" instead of a streaming value. Use
+    /// if only the instantanouse value is required.
+    /// </para>
+    /// <para>
+    /// Peek and poke are similar to `property` members. Unlike streaming,
+    /// peek and poke are reliable operations.
+    /// </para>
+    /// <para>
+    /// Throws ValueNotSetException if OutValue is not valid.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="ts">[out] The timestamp of the current OutValue</param>
+    /// <returns>The current OutValue</returns>
     public T PeekOutValue(out TimeSpec ts)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2298,6 +1522,23 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Poke the OutValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Pokes the OutValue using a "request" instead of a streaming value. Use
+    /// to update the OutValue if the value is updated infrequently.
+    /// </para>
+    /// <para>
+    /// Peek and poke are similar to `property` members. Unlike streaming,
+    /// peek and poke are reliable operations.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="value">The new OutValue</param>
     public void PokeOutValue(T value)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2316,6 +1557,19 @@ public class Wire<T>
         {}
     }
 
+    /// <summary>
+    /// Asynchronously peek the current InValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Same as PeekInValue(), but returns asynchronously.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>Task that upon completion returns the packet and timespec</returns>
     public async Task<Tuple<T, TimeSpec>> AsyncPeekInValue(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2328,6 +1582,19 @@ public class Wire<T>
         return await h.Task;
     }
 
+    /// <summary>
+    /// Asynchronously peek the current OutValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Same as PeekOutValue(), but async.
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>Task that upon completion returns the packet and timespec</returns>
     public async Task<Tuple<T, TimeSpec>> AsyncPeekOutValue(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2340,6 +1607,19 @@ public class Wire<T>
         return await h.Task;
     }
 
+    /// <summary>
+    /// Asynchronously poke the OutValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Same as PokeOutValue(), but async
+    /// </para>
+    /// <para>
+    /// Only valid on clients. Will throw InvalidOperationException on the service side.
+    /// </para>
+    /// </remarks>
+    /// <param name="value">The new OutValue</param>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
     public async Task AsyncPokeOutValue(T value, int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
         WrappedWireClient c = innerwire as WrappedWireClient;
@@ -2393,6 +1673,36 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Set the PeekInValue callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Peek and poke operations are used when a streaming connection of the most recent value
+    /// is not required. Clients initiate peek and poke operations using PeekInValue(), PeekOutValue(),
+    /// PokeOutValue(), or their asynchronous equivalents. Services receive the peek and poke
+    /// requests through callbacks.
+    /// </para>
+    /// <para>
+    /// PeekInValueCallback configures the service callback for PeekInValue() requests.
+    /// </para>
+    /// <para>
+    /// The specified callback function should have the following signature:
+    /// </para>
+    /// <para>
+    /// T peek_invalue_callback(uint32 client_endpoint);
+    /// </para>
+    /// <para> 
+    /// The function receives the client endpoint ID, and returns the current InValue.
+    /// </para>
+    /// <para>
+    /// Note: Callback is configured automatically by WireBroadcaster or
+    /// WireUnicastReceiver
+    /// </para>
+    /// <para>
+    /// Only valid for services. Will throw InvalidOperationException on the client side.
+    /// </para>
+    /// </remarks>
     public Func<uint, T> PeekInValueCallback
     {
         get {
@@ -2405,6 +1715,36 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Set the PeekOutValue callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Peek and poke operations are used when a streaming connection of the most recent value
+    /// is not required. Clients initiate peek and poke operations using PeekInValue(), PeekOutValue(),
+    /// PokeOutValue(), or their asynchronous equivalents. Services receive the peek and poke
+    /// requests through callbacks.
+    /// </para>
+    /// <para>
+    /// PeekOutValueCallback configures the service callback for PeekOutValue() requests.
+    /// </para>
+    /// <para>
+    /// The specified callback function should have the following signature:
+    /// </para>
+    /// <para>
+    ///     T peek_outvalue_callback(uint32 client_endpoint);
+    /// </para>
+    /// <para>
+    /// The function receives the client endpoint ID, and returns the current OutValue.
+    /// </para>
+    /// <para>
+    /// Note: Callback is configured automatically by WireBroadcaster or
+    /// WireUnicastReceiver
+    /// </para>
+    /// <para>
+    /// Only valid for services. Will throw InvalidOperationException on the client side.
+    /// </para>
+    /// </remarks>
     public Func<uint, T> PeekOutValueCallback
     {
         get {
@@ -2417,6 +1757,37 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Set the PokeOutValue callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Peek and poke operations are used when a streaming connection of the most recent value
+    /// is not required. Clients initiate peek and poke operations using PeekInValue(), PeekOutValue(),
+    /// PokeOutValue(), or their asynchronous equivalents. Services receive the peek and poke
+    /// requests through callbacks.
+    /// </para>
+    /// <para>
+    /// PokeOutValueCallback configures the service callback for PokeOutValue() requests.
+    /// </para>
+    /// <para>
+    /// The specified callback function should have the following signature:
+    /// </para>
+    /// <para>
+    ///     void poke_outvalue_callback(const T& value, const TimeSpec& timestamp, uint32 client_endpoint);
+    /// </para>
+    /// <para>
+    /// The function receives the new out value, the new out value timestamp in the client's clock,
+    /// and the client endpoint ID.
+    /// </para>
+    /// <para>
+    /// Note: Callback is configured automatically by WireBroadcaster or
+    /// WireUnicastReceiver
+    /// </para>
+    /// <para>
+    /// Only valid for services. Will throw InvalidOperationException on the client side.
+    /// </para>
+    /// </remarks>
     public Action<T, TimeSpec, uint> PokeOutValueCallback
     {
         get {
@@ -2429,6 +1800,38 @@ public class Wire<T>
         }
     }
 
+    /// <summary>
+    /// Wire connection used to transmit "most recent" values
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Wire connections are used to transmit "most recent" values between connected
+    /// wire members. See Wire for more information on wire members.
+    /// </para>
+    /// <para>
+    /// Wire connections are created by clients using the Wire::Connect() or Wire::AsyncConnect()
+    /// functions. Services receive incoming wire connection requests through a
+    /// callback function specified using the Wire::SetWireConnectCallback() function. Services
+    /// may also use the WireBroadcaster class to automate managing wire connection lifecycles and
+    /// sending values to all connected clients, or use WireUnicastReceiver to receive an incoming
+    /// value from the most recently connected client.
+    /// </para>
+    /// <para>
+    /// Wire connections are used to transmit "most recent" values between clients and services. Connection
+    /// the wire creates a connection pair, one in the client, and one in the service. Each wire connection
+    /// object has an InValue and an OutValue. Setting the OutValue of one will cause the specified value to
+    /// be transmitted to the InValue of the peer. See Wire for more information.
+    /// </para>
+    /// <para>
+    /// Values can optionally be specified to have a finite lifespan using SetInValueLifespan() and
+    /// OutValueLifespan. Lifespans can be used to prevent using old values that have
+    /// not been recently updated.
+    /// </para>
+    /// <para>
+    /// This class is instantiated by the Wire class. It should not be instantiated
+    /// by the user.
+    /// </para>
+    /// </remarks>
     public class WireConnection
     {
         WrappedWireConnection innerwire;
@@ -2459,6 +1862,14 @@ public class Wire<T>
             {}
         }
 
+        /// <summary>
+        /// The direction of the wire
+        /// </summary>
+        /// <remarks>
+        /// Wires may be declared*readonly* or*writeonly* in the service definition file. (If neither
+        /// is specified, the wire is assumed to be full duplex.)*readonly* wire may only send out values from
+        /// service to client.*writeonly* wires may only send out values from client to service.
+        /// </remarks>
         public MemberDefinition_Direction Direction
         {
             get {
@@ -2466,6 +1877,13 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get the Robot Raconteur node Endpoint ID
+        /// </summary>
+        /// <remarks>
+        /// Returns the endpoint associated with the ClientContext or ServerEndpoint
+        /// associated with the wire connection.
+        /// </remarks>
         public uint Endpoint
         {
             get {
@@ -2473,6 +1891,14 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get the current InValue
+        /// </summary>
+        /// <remarks>
+        /// Gets the current InValue that was transmitted from the peer. Throws
+        /// ValueNotSetException if no value has been received, or the most
+        /// recent value lifespan has expired.
+        /// </remarks>
         public T InValue
         {
             get {
@@ -2483,6 +1909,21 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get or set the current OutValue
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Gets the current OutValue that was transmitted to the peer. Throws
+        /// ValueNotSetException if no value has been received, or the most
+        /// recent value lifespan has expired.
+        /// </para>
+        /// <para> 
+        /// Setting the OutValue for the wire connection. The specified value will be
+        /// transmitted to the peer, and will become the peers InValue. The transmission
+        /// is unreliable, meaning that values may be dropped if newer values arrive.
+        /// </para>
+        /// </remarks>
         public T OutValue
         {
             get {
@@ -2506,6 +1947,12 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get the timestamp of the last received value
+        /// </summary>
+        /// <remarks>
+        /// Returns the timestamp of the value in the///senders* clock
+        /// </remarks>
         public TimeSpec LastValueReceivedTime
         {
             get {
@@ -2513,6 +1960,12 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get the timestamp of the last sent value
+        /// </summary>
+        /// <remarks>
+        /// Returns the timestamp of the last sent value in the///local* clock
+        /// </remarks>
         public TimeSpec LastValueSentTime
         {
             get {
@@ -2520,6 +1973,16 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Try getting the InValue, returning true on success or false on failure
+        /// </summary>
+        /// <remarks>
+        /// Get the current InValue and InValue timestamp. Return true or false on
+        /// success or failure instead of throwing exception.
+        /// </remarks>
+        /// <param name="value">[out] The current InValue</param>
+        /// <param name="ts">[out] The current InValue timestamp in the senders clock</param>
+        /// <returns>true if the value is valid, otherwise false</returns>
         bool TryGetInValue(out T value, out TimeSpec ts)
         {
             value = default(T);
@@ -2535,6 +1998,16 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Try getting the OutValue, returning true on success or false on failure
+        /// </summary>
+        /// <remarks>
+        /// Get the current OutValue and OutValue timestamp. Return true or false on
+        /// success and failure instead of throwing exception.
+        /// </remarks>
+        /// <param name="value">[out] The current OutValue</param>
+        /// <param name="ts">[out] The current OutValue timestamp in the local clock</param>
+        /// <returns>true if the value is valid, otherwise false</returns>
         bool TryGetOutValue(out T value, out TimeSpec ts)
         {
             value = default(T);
@@ -2550,11 +2023,25 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Close the wire connection
+        /// </summary>
+        /// <remarks>
+        /// Close the wire connection. Blocks until close complete. The peer wire connection
+        /// is destroyed automatically.
+        /// </remarks>
         public void Close()
         {
             innerwire.Close();
         }
 
+        /// <summary>
+        /// Asynchronously close the wire connection
+        /// </summary>
+        /// <remarks>
+        /// Same as Close() but async
+        /// </remarks>
+        /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
         public async Task AsyncClose(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
         {
             AsyncVoidReturnDirectorImpl h = new AsyncVoidReturnDirectorImpl();
@@ -2563,10 +2050,30 @@ public class Wire<T>
             await h.Task;
         }
 
+        /// <summary>
+        /// Event invoked when the InValue is changed
+        /// </summary>
+        /// <remarks>
+        /// Callback function must accept three arguments, receiving the WireConnection that
+        /// received a packet, the new value, and the value's TimeSpec timestamp
+        /// </remarks>
         public event Action<Wire<T>.WireConnection, T, TimeSpec> WireValueChanged;
 
         private Action<WireConnection> close_callback;
 
+
+        /// <summary>
+        /// Get or set the connection closed callback function
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Sets a function to invoke when the wire connection has been closed.
+        /// </para>
+        /// <para>
+        /// Callback function must accept one argument, receiving the WireConnectionPtr<T> that
+        /// was closed.
+        /// </para>
+        /// </remarks>
         public Action<WireConnection> WireCloseCallback
         {
             get {
@@ -2577,6 +2084,13 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get if the InValue is valid
+        /// </summary>
+        /// <remarks>
+        /// The InValue is valid if a value has been received and
+        /// the value has not expired
+        /// </remarks>
         public bool InValueValid
         {
             get {
@@ -2584,6 +2098,13 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get if the OutValue is valid
+        /// </summary>
+        /// <remarks>
+        /// The OutValue is valid if a value has been
+        /// set using OutValue
+        /// </remarks>
         public bool OutValueValid
         {
             get {
@@ -2591,6 +2112,14 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get or set whether wire connection should ignore incoming values
+        /// </summary>
+        /// <remarks>
+        /// Wire connections may optionally desire to ignore incoming values. This is useful if the connection
+        /// is only being used to send out values, and received values may create a potential memory . If ignore is true,
+        /// incoming values will be discarded.
+        /// </remarks>
         public bool IgnoreInValue
         {
             get {
@@ -2601,6 +2130,24 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get or set the lifespan of InValue
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// InValue may optionally have a finite lifespan specified in milliseconds. Once
+        /// the lifespan after reception has expired, the InValue is cleared and becomes invalid.
+        /// Attempts to access InValue will result in ValueNotSetException.
+        /// </para>
+        /// <para>
+        /// InValue lifespans may be used to avoid using a stale value received by the wire. If
+        /// the lifespan is not set, the wire will continue to return the last received value, even
+        /// if the value is old.
+        /// </para>
+        /// <para> 
+        /// The lifespan in millisecond, or RR_VALUE_LIFESPAN_INFINITE for infinite lifespan
+        /// </para>
+        /// </remarks>
         public int InValueLifespan
         {
             get {
@@ -2611,6 +2158,24 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Get or set the lifespan of OutValue
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// OutValue may optionally have a finite lifespan specified in milliseconds. Once
+        /// the lifespan after sending has expired, the OutValue is cleared and becomes invalid.
+        /// Attempts to access OutValue will result in ValueNotSetException.
+        /// </para>
+        /// <para>
+        /// OutValue lifespans may be used to avoid using a stale value sent by the wire. If
+        /// the lifespan is not set, the wire will continue to return the last sent value, even
+        /// if the value is old.
+        /// </para>
+        /// <para> 
+        /// The lifespan in millisecond, or RR_VALUE_LIFESPAN_INFINITE for infinite lifespan
+        /// </para>
+        /// </remarks>
         public int OutValueLifespan
         {
             get {
@@ -2621,11 +2186,31 @@ public class Wire<T>
             }
         }
 
+        /// <summary>
+        /// Waits for InValue to be valid
+        /// </summary>
+        /// <remarks>
+        /// Blocks the current thread until InValue is valid,
+        /// with an optional timeout. Returns true if InValue is valid,
+        /// or false if timeout occurred.
+        /// </remarks>
+        /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+        /// <returns>true if InValue is valid, otherwise false</returns>
         public bool WaitInValueValid(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
         {
             return innerwire.WaitInValueValid(timeout);
         }
 
+        /// <summary>
+        /// Waits for OutValue to be valid
+        /// </summary>
+        /// <remarks>
+        /// Blocks the current thread until OutValue is valid,
+        /// with an optional timeout. Returns true if OutValue is valid,
+        /// or false if timeout occurred.
+        /// </remarks>
+        /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+        /// <returns>true if InValue is valid, otherwise false</returns>
         public bool WaitOutValueValid(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
         {
             return innerwire.WaitOutValueValid(timeout);
@@ -2698,6 +2283,36 @@ public class Wire<T>
     }
 }
 
+/// <summary>
+/// Broadcaster to send values to all connected clients
+/// </summary>
+/// <remarks>
+/// <para>
+/// WireBroadcaster is used by services to send values to all
+/// connected client endpoints. It attaches to the wire on the service
+/// side, and manages the lifecycle of connections. WireBroadcaster
+/// should only we used with wires that are declared*readonly*, since
+/// it has no provisions for receiving incoming values from clients.
+/// </para>
+/// <para>
+/// WireBroadcaster is initialized by the user, or by default implementation
+/// classes generated by RobotRaconteurGen (*_default_impl). Default
+/// implementation classes will automatically instantiate broadcasters for
+/// wires marked*readonly*. If default implementation classes are
+/// not used, the broadcaster must be instantiated manually. It is recommended this
+/// be done using the IRRServiceObject interface in the overridden
+/// IRRServiceObject::RRServiceObjectInit() function. This function is called after
+/// the wires have been instantiated by the service.
+/// </para>
+/// <para>
+/// Set the OutValue property to broadcast values to all connected clients.
+/// </para>
+/// <para>
+/// The rate that packets are sent can be regulated using a callback function configured
+/// with the Predicate property, or using the BroadcastDownsampler class.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The value data type</typeparam>
 public class WireBroadcaster<T>
 {
     class WrappedWireBroadcasterPredicateDirectorNET : WrappedWireBroadcasterPredicateDirector
@@ -2738,6 +2353,11 @@ public class WireBroadcaster<T>
     protected internal WrappedWireBroadcaster innerwire;
     protected TypeDefinition type;
 
+    /// <summary>
+    /// Construct a new WireBroadcaster
+    /// </summary>
+    /// <param name="wire">The wire to use for broadcasting. Must be a wire from a service object.
+    /// Specifying a client wire will result in an exception.</param>
     public WireBroadcaster(Wire<T> wire)
     {
         this.wire = wire;
@@ -2746,6 +2366,9 @@ public class WireBroadcaster<T>
         this.type = ((WrappedWireServer)wire.innerwire).Type;
     }
 
+    /// <summary>
+    /// Get the assosciated wire
+    /// </summary>
     public Wire<T> Wire
     {
         get {
@@ -2753,6 +2376,20 @@ public class WireBroadcaster<T>
         }
     }
 
+    /// <summary>
+    /// Set the OutValue for all connections
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Sets the OutValue for all connections. This will transmit the value
+    /// to all connected clients using packets. The value will become the clients'
+    /// InValue.
+    /// </para>
+    /// <para>
+    /// The value will be returned when clients call Wire.PeekInValue() or
+    /// Wire.AsyncPeekInValue()
+    /// </para>
+    /// </remarks>
     public T OutValue
     {
         set {
@@ -2768,6 +2405,9 @@ public class WireBroadcaster<T>
         }
     }
 
+    /// <summary>
+    /// Get the current number of active wire connections
+    /// </summary>
     public uint ActiveWireConnectionCount
     {
         get {
@@ -2775,6 +2415,27 @@ public class WireBroadcaster<T>
         }
     }
 
+    /// <summary>
+    /// Set the predicate callback function
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// A predicate is optionally used to regulate when values are sent to clients. This is used by the
+    /// BroadcastDownsampler to regulate update rates of values sent to clients.
+    /// </para>
+    /// <para>
+    /// The predicate callback is invoked before the broadcaster sets the OutValue of a connection. If the predicate
+    /// returns true, the OutValue packet will be sent. If it is false, the OutValue packet will not be sent to that
+    /// endpoint. The predicate callback must have the following signature:
+    /// </para>
+    /// <para>
+    ///     bool broadcaster_predicate(WireBroadcaster broadcaster, uint client_endpoint);
+    /// </para>
+    /// <para>
+    /// It receives the broadcaster and the client endpoint ID. It returns true to send the OutValue packet,
+    /// or false to not send the OutValue packet.
+    /// </para>
+    /// </remarks>
     public Func<uint, bool> Predicate
     {
         set {
@@ -2784,6 +2445,24 @@ public class WireBroadcaster<T>
         }
     }
 
+    /// <summary>
+    /// Get or set the lifespan of OutValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// OutValue may optionally have a finite lifespan specified in milliseconds. Once
+    /// the lifespan after sending has expired, the OutValue is cleared and becomes invalid.
+    /// Attempts to access OutValue will result in ValueNotSetException.
+    /// </para>
+    /// <para>
+    /// OutValue lifespans may be used to avoid using a stale value sent by the wire. If
+    /// the lifespan is not set, the wire will continue to return the last sent value, even
+    /// if the value is old.
+    /// </para>
+    /// <para> 
+    /// The lifespan in millisecond, or RR_VALUE_LIFESPAN_INFINITE for infinite lifespan
+    /// </para>
+    /// </remarks>
     public int OutValueLifespan
     {
         get {
@@ -2793,6 +2472,7 @@ public class WireBroadcaster<T>
             innerwire.SetOutValueLifespan(value);
         }
     }
+
 
     public Func<uint, T> PeekInValueCallback
     {
@@ -2807,12 +2487,52 @@ public class WireBroadcaster<T>
     }
 }
 
+/// <summary>
+/// Receive the InValue from the most recent connection
+/// </summary>
+/// <remarks>
+/// <para>
+/// WireUnicastReceiver is used by services to receive a value from a single client.
+/// When a client sets its OutValue, this value is transmitted to the service using
+/// packets, and becomes the service's InValue for that connection. Service wires
+/// can have multiple active clients, so the service needs to choose which connection
+/// is "active". The WireUnicastReceiver selects the "most recent" connection, and
+/// returns that connection's InValue. Any existing connections are closed.
+/// WireUnicastReceiver should only be used with wires that are declared*writeonly*.
+/// It is recommended that object locks be used to protect from concurrent
+/// access when unicast receivers are used.
+/// </para>
+/// <para>
+/// WireUnicastReceiver is initialized by the user, or by default implementation
+/// classes generated by RobotRaconteurGen (*_default_impl). Default
+/// implementation classes will automatically instantiate unicast receivers for
+/// wires marked*writeonly*. If default implementation classes are
+/// not used, the unicast receiver must be instantiated manually. It is recommended this
+/// be done using the IRRServiceObject interface in the overridden
+/// IRRServiceObject::RRServiceObjectInit() function. This function is called after
+/// the wires have been instantiated by the service.
+/// </para>
+/// <para>
+/// The current InValue is received using GetInValue() or TryGetInValue(). The
+/// InValueChanged signal can be used to monitor for changes to the InValue.
+/// </para>
+/// <para>
+/// Clients may also use PokeOutValue() or AsyncPokeOutValue() to update the
+/// unicast receiver's value.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The value type</typeparam>
 public class WireUnicastReceiver<T>
 {
     protected Wire<T> wire;
     protected WrappedWireUnicastReceiver innerwire;
     protected TypeDefinition type;
 
+    /// <summary>
+    /// Construct a new WireUnicastReceiverBase
+    /// </summary>
+    /// <param name="wire">The wire to use for broadcasting. Must be a wire from a service object.
+    /// Specifying a client wire will result in an exception.</param>
     public WireUnicastReceiver(Wire<T> wire)
     {
         this.wire = wire;
@@ -2825,6 +2545,9 @@ public class WireUnicastReceiver<T>
         this.innerwire.AddInValueChangedListener(cb, id);
     }
 
+    /// <summary>
+    /// Get the associated wire
+    /// </summary>
     public Wire<T> Wire
     {
         get {
@@ -2832,6 +2555,18 @@ public class WireUnicastReceiver<T>
         }
     }
 
+
+    /// <summary>
+    /// Get the current InValue
+    /// </summary>
+    /// <remarks>
+    /// Gets the current InValue that was received from the active connection.
+    /// Throws ValueNotSetException if no value has been received, or
+    /// the most recent value lifespan has expired.
+    /// </remarks>
+    /// <param name="ts">[out] The current InValue timestamp</param>
+    /// <param name="ep">[out] The client endpoint ID of the InValue</param>
+    /// <returns>The current InValue</returns>    
     public T GetInValue(out TimeSpec ts, out uint ep)
     {
         ts = new TimeSpec();
@@ -2839,6 +2574,18 @@ public class WireUnicastReceiver<T>
         return RobotRaconteurNode.s.UnpackAnyType<T>(m);
     }
 
+    /// <summary>
+    /// Try getting the current InValue, returning true on success or false on failure
+    /// </summary>
+    /// <remarks>
+    /// Gets the current InValue, its timestamp, and the client endpoint ID. Returns true if
+    /// value is valid, or false if value is invalid. Value will be invalid if no value has
+    /// been received, or the value lifespan has expired.
+    /// </remarks>
+    /// <param name="value">[out] The current InValue</param>
+    /// <param name="time">[out] The current InValue timestamp</param>
+    /// <param name="client">[out] The client endpoint ID of the InValue</param>
+    /// <returns>true if value is valid, otherwise false</returns>
     public bool TryGetInValue(out T value, out TimeSpec time, out uint client)
     {
         var m = new WrappedService_typed_packet();
@@ -2893,8 +2640,33 @@ public class WireUnicastReceiver<T>
         }
     }
 
+    /// <summary>
+    /// Event fired when InValue has changed.
+    /// </summary>
+    /// <remarks>
+    /// Callback function must accept three arguments, receiving the new value,
+    /// value's TimeSpec timestamp, and the client endpoint ID.
+    /// </remarks>
     public event Action<T, TimeSpec, uint> InValueChanged;
 
+    /// <summary>
+    /// Get or set the lifespan of InValue
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// InValue may optionally have a finite lifespan specified in milliseconds. Once
+    /// the lifespan after reception has expired, the InValue is cleared and becomes invalid.
+    /// Attempts to access InValue will result in ValueNotSetException.
+    /// </para>
+    /// <para>
+    /// InValue lifespans may be used to avoid using a stale value received by the wire. If
+    /// the lifespan is not set, the wire will continue to return the last received value, even
+    /// if the value is old.
+    /// </para>
+    /// <para> 
+    /// The lifespan in millisecond, or RR_VALUE_LIFESPAN_INFINITE for infinite lifespan
+    /// </para>
+    /// </remarks>
     public int InValueLifespan
     {
         get {
@@ -2906,12 +2678,33 @@ public class WireUnicastReceiver<T>
     }
 }
 
+/// <summary>
+/// Service object monitor lock notification
+/// </summary>
+/// <remarks>
+/// Service objects must implement IRobotRaconteurMonitorObject for
+/// monitor locking to function. Services call RobotRaconteurMonitorEnter()
+/// with an optional timeout to request the lock, and call RobotRaconteurMonitorExit()
+/// to release the monitor lock. RobotRaconteurMonitorEnter() should block
+/// until a thread-exclusive lock can be established.
+/// </remarks>
 public interface IRobotRaconteurMonitorObject
 {
+    /// <summary>
+    /// Request a thread-exclusive lock without timeout. May block until lock can be established
+    /// </summary>
     void RobotRaconteurMonitorEnter();
 
+    /// <summary>
+    /// Request a thread-exclusive lock with timeout. May block until lock can be established, 
+    /// up to the specified timeout.
+    /// </summary>
+    /// <param name="timeout">Lock request timeout in milliseconds</param>
     void RobotRaconteurMonitorEnter(int timeout);
 
+    /// <summary>
+    /// Release the thread-exclusive monitor lock
+    /// </summary>
     void RobotRaconteurMonitorExit();
 }
 
@@ -3022,8 +2815,29 @@ public partial class ServiceSecurityPolicy
     }
 }
 
+/// <summary>
+/// Interface for custom user authenticators
+/// </summary>
 public interface IUserAuthenticator
 {
+    /// <summary>
+    /// Authenticate a user using username and credentials
+    /// </summary>
+    /// <remarks>
+    /// The service will call AuthenticateUser when clients attempt to authenticate. The
+    /// username and credentials are passed to RobotRaconteurNode.ConnectService() or
+    /// RobotRaconteurNode.AsyncConnectService(). The authenticator must analyze these
+    /// values to decide how to authenticate the user.
+    ///
+    /// On successful authentication, return a populated AuthenticatedUser object. On
+    /// failure, throw an exception.
+    ///
+    /// See \ref security for more information.
+    /// </remarks>
+    /// <param name="username">The username provided by the client</param>
+    /// <param name="credentials">The credentials provided by the client</param>
+    /// <param name="context">The context of the service requesting authentication</param>
+    /// <returns>An authenticated user object</returns>
     AuthenticatedUser AuthenticateUser(string username, Dictionary<string, object> credentials, ServerContext context);
 }
 
@@ -3040,6 +2854,9 @@ public partial class NativeUserAuthenticator
 
 public partial class AuthenticatedUser
 {
+    /// <summary>
+    /// The authenticated username
+    /// </summary>
     public string Username
     {
         get {
@@ -3047,6 +2864,9 @@ public partial class AuthenticatedUser
         }
     }
 
+    /// <summary>
+    /// The user privileges
+    /// </summary>
     public string[] Privileges
     {
         get {
@@ -3057,6 +2877,9 @@ public partial class AuthenticatedUser
         }
     }
 
+    /// <summary>
+    /// The user properties
+    /// </summary>
     public string[] Properties
     {
         get {
@@ -3067,6 +2890,9 @@ public partial class AuthenticatedUser
         }
     }
 
+    /// <summary>
+    /// The user login time
+    /// </summary>
     public DateTime LoginTime
     {
         get {
@@ -3074,11 +2900,47 @@ public partial class AuthenticatedUser
         }
     }
 
+    /// <summary>
+    /// The user last access time
+    /// </summary>
     public DateTime LastAccessTime
     {
         get {
             return _GetLastAccessTime();
         }
+    }
+}
+
+/// <summary>
+///  Class for automatic broadcast downsampler stepping
+/// </summary>
+/// <remarks>
+/// Helper class to automate BroadcastDownsampler stepping.
+/// Calls BroadcastDownsampler.BeginStep() on construction,
+/// and BroadcastDownsampler.EndStep() when disposed.
+/// </remarks>
+public class BroadcastDownsamplerStep : IDisposable
+{
+    protected BroadcastDownsampler downsampler;
+
+    /// <summary>
+    /// Construct a BroadcastDownsampler
+    /// </summary>
+    /// <remarks>
+    /// Calls BroadcastDownsampler.BeginStep() on downsampler.
+    /// Calls BroadcastDownsampler.EndStep() on downsampler
+    /// when destroyed.
+    /// </remarks>
+    /// <param name="downsampler">The downsampler to step</param>
+    public BroadcastDownsamplerStep(BroadcastDownsampler downsampler)
+    {
+        this.downsampler = downsampler;
+        downsampler.BeginStep();
+    }
+
+    public void Dispose()
+    {
+        downsampler.EndStep();
     }
 }
