+++ RobotRaconteurNET/RobotRaconteurNET/Subscription.cs
@@ -25,12 +25,30 @@ namespace RobotRaconteur
 {
 // Subscriptions
 
+/// <summary>
+/// ClientID for use with ServiceSubscription
+/// </summary>
+/// <remarks>
+/// The ServiceSubscriptionClientID stores the NodeID
+/// and ServiceName of a connected service.
+/// </remarks>
 public partial class ServiceSubscriptionClientID
 {
 
+    /// <summary>
+    /// The NodeID of the connected service
+    /// </summary>
     public NodeID NodeID;
+    /// <summary>
+    /// The ServiceName of the connected service
+    /// </summary>
     public string ServiceName;
 
+    /// <summary>
+    /// Construct a ServiceSubscriptionClientID
+    /// </summary>
+    /// <param name="node_id">The NodeID</param>
+    /// <param name="service_name">The Service Name</param>
     public ServiceSubscriptionClientID(NodeID node_id, string service_name)
     {
         this.NodeID = node_id;
@@ -43,6 +61,9 @@ public partial class ServiceSubscriptionClientID
         this.ServiceName = id1.ServiceName;
     }
 
+    /// <summary>
+    /// Equality operator
+    /// </summary>
     public static bool operator ==(ServiceSubscriptionClientID a, ServiceSubscriptionClientID b)
     {
         if (Object.Equals(a, null) && Object.Equals(b, null))
@@ -52,6 +73,9 @@ public partial class ServiceSubscriptionClientID
         return (a.NodeID == b.NodeID) && (a.ServiceName == b.ServiceName);
     }
 
+    /// <summary>
+    /// Inequality operator
+    /// </summary>
     public static bool operator !=(ServiceSubscriptionClientID a, ServiceSubscriptionClientID b)
     {
         return !(a == b);
@@ -70,23 +94,83 @@ public partial class ServiceSubscriptionClientID
     }
 }
 
+/// <summary>
+/// Subscription filter node information
+/// </summary>
+/// <remarks>
+/// Specify a node by NodeID and/or NodeName. Also allows specifying
+/// username and password.
+///
+/// When using username and credentials, secure transports and specified NodeID should
+/// be used. Using username and credentials without a transport that verifies the
+/// NodeID could result in credentials being leaked.
+/// </remarks>
 public class ServiceSubscriptionFilterNode
 {
+    /// <summary>
+    /// The NodeID to match. All zero NodeID will match any NodeID.
+    /// </summary>
     public NodeID NodeID = RobotRaconteur.NodeID.GetAny();
+    /// <summary>
+    /// The NodeName to match. Empty or null NodeName will match any NodeName.
+    /// </summary>
     public string NodeName;
+    /// <summary>
+    /// The username to use for authentication. Should only be used with secure transports and verified NodeID
+    /// </summary>
     public string Username;
+    /// <summary>
+    /// The credentials to use for authentication. Should only be used with secure transports and verified NodeID
+    /// </summary>
     public Dictionary<string, object> Credentials;
 }
 
+/// <summary>
+/// Subscription filter
+/// </summary>
+/// <remarks>
+/// The subscription filter is used with RobotRaconteurNode.SubscribeServiceByType() and
+/// RobotRaconteurNode::SubscribeServiceInfo2() to decide which services should
+/// be connected. Detected services that match the service type are checked against
+/// the filter before connecting.
+/// </remarks>
 public class ServiceSubscriptionFilter
 {
+    /// <summary>
+    /// Vector of nodes that should be connected. Empty means match any node.
+    /// </summary>
     public ServiceSubscriptionFilterNode[] Nodes;
+    /// <summary>
+    /// Vector service names that should be connected. Empty means match any service name.
+    /// </summary>
     public string[] ServiceNames;
+    /// <summary>
+    /// Vector of transport schemes. Empty means match any transport scheme.
+    /// </summary>
     public string[] TransportSchemes;
+    /// <summary>
+    /// A user specified predicate function. If nullptr, the predicate is not checked.
+    /// </summary>
     public Func<ServiceInfo2, bool> Predicate;
+    /// <summary>
+    /// The maximum number of connections the subscription will create. Zero means unlimited connections. 
+    /// </summary>
     public uint MaxConnections;
 }
 
+/// <summary>
+/// Subscription for information about detected services
+/// </summary>
+/// <remarks>
+/// <para>
+/// Created using RobotRaconteurNode::SubscribeServiceInfo2()
+/// </para>
+/// <para>
+/// The ServiceInfo2Subscription class is used to track services with a specific service type as they are
+/// detected on the local network and when they are lost. The currently detected services can also
+/// be retrieved. The service information is returned using the ServiceInfo2 structure.
+/// </para>
+/// </remarks>
 public class ServiceInfo2Subscription
 {
     class WrappedServiceInfo2SubscriptionDirectorNET : WrappedServiceInfo2SubscriptionDirector
@@ -145,11 +229,27 @@ public class ServiceInfo2Subscription
         subscription.SetRRDirector(director, id);
     }
 
+    /// <summary>
+    /// Close the subscription
+    /// </summary>
+    /// <remarks>
+    /// Closes the subscription. Subscriptions are automatically closed when the node is shut down.
+    /// </remarks>
     public void Close()
     {
         _subscription.Close();
     }
 
+    /// <summary>
+    /// Returns a dictionary of detected services.
+    /// </summary>
+    /// <remarks>
+    /// The returned dictionary contains the detected nodes as ServiceInfo2. The map
+    /// is keyed with ServiceSubscriptionClientID.
+    ///
+    /// This function does not block.
+    /// </remarks>
+    /// <returns>The detected services</returns>
     public Dictionary<ServiceSubscriptionClientID, ServiceInfo2> GetDetectedServiceInfo2()
     {
         var o = new Dictionary<ServiceSubscriptionClientID, ServiceInfo2>();
@@ -165,10 +265,42 @@ public class ServiceInfo2Subscription
         return o;
     }
 
+    /// <summary>
+    /// Listener event that is invoked when a service is detected
+    /// </summary>
     public event Action<ServiceInfo2Subscription, ServiceSubscriptionClientID, ServiceInfo2> ServiceDetected;
+    /// <summary>
+    /// Listener event that is invoked when a service is lost
+    /// </summary>
     public event Action<ServiceInfo2Subscription, ServiceSubscriptionClientID, ServiceInfo2> ServiceLost;
 }
 
+/// <summary>
+/// Subscription that automatically connects services and manages lifecycle of connected services
+/// </summary>
+/// <remarks>
+/// <para>
+/// Created using RobotRaconteurNode.SubscribeService() or RobotRaconteurNode.SubscribeServiceByType(). The
+/// ServiceSubscription class is used to automatically create and manage connections based on connection criteria.
+/// RobotRaconteurNode.SubscribeService() is used to create a robust connection to a service with a specific URL.
+/// RobotRaconteurNode.SubscribeServiceByType() is used to connect to services with a specified type, filtered with a
+/// ServiceSubscriptionFilter. Subscriptions will create connections to matching services, and will retry the connection
+/// if it fails or the connection is lost. This behavior allows subscriptions to be used to create robust connections.
+/// The retry delay for connections can be modified using ConnectRetryDelay.
+/// </para>
+/// <para>
+/// The currently connected clients can be retrieved using the GetConnectedClients() function. A single "default client"
+/// can be retrieved using the GetDefaultClient() function or TryGetDefaultClient() functions. Listeners for client
+/// connect and disconnect events can be added  using the AddClientConnectListener() and AddClientDisconnectListener()
+/// functions. If the user wants to claim a client, the ClaimClient() and ReleaseClient() functions will be used. Claimed
+/// clients will no longer have their lifecycle managed by the subscription.
+/// </para>
+/// <para>
+/// Subscriptions can be used to create `pipe` and `wire` subscriptions. These member subscriptions aggregate
+/// the packets and values being received from all services. They can also act as a "reverse broadcaster" to
+/// send packets and values to all services that are actively connected. See PipeSubscription and WireSubscription.
+/// </para>
+/// </remarks>
 public class ServiceSubscription
 {
 
@@ -296,6 +428,27 @@ public class ServiceSubscription
         subscription.SetRRDirector(director, id);
     }
 
+    /// <summary>
+    /// Returns a dictionary of connected clients
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// The returned dictionary contains the connect clients. The map
+    /// is keyed with ServiceSubscriptionClientID.
+    /// </para>
+    /// <para>
+    /// Clients must be cast to a type, similar to the client returned by
+    /// RobotRaconteurNode.ConnectService().
+    /// </para>
+    /// <para>
+    /// Clients can be "claimed" using ClaimClient(). Once claimed, the subscription
+    /// will stop managing the lifecycle of the client.
+    /// </para>
+    /// <para>
+    /// This function does not block.
+    /// </para>
+    /// </remarks>
+    /// <returns>The detected services.</returns>
     public Dictionary<ServiceSubscriptionClientID, object> GetConnectedClients()
     {
         var o = new Dictionary<ServiceSubscriptionClientID, object>();
@@ -311,11 +464,26 @@ public class ServiceSubscription
         return o;
     }
 
+    /// <summary>
+    /// Close the subscription
+    /// </summary>
+    /// <remarks>
+    /// Close the subscription. Subscriptions are automatically closed when the node is shut down.
+    /// </remarks>
     public void Close()
     {
         _subscription.Close();
     }
 
+    /// <summary>
+    /// Claim a client that was connected by the subscription
+    /// </summary>
+    /// <remarks>
+    /// The subscription class will automatically manage the lifecycle of the connected clients. The clients
+    /// will be automatically disconnected and/or reconnected as necessary. If the user wants to disable
+    /// this behavior for a specific client connection, the client connection can be "claimed".
+    /// </remarks>
+    /// <param name="client">The client to be claimed</param>
     public void ClaimClient(object client)
     {
         ServiceStub s = client as ServiceStub;
@@ -325,6 +493,13 @@ public class ServiceSubscription
         _subscription.ClaimClient(s.rr_innerstub);
     }
 
+    /// <summary>
+    /// Release a client previously clamed with ClaimClient()
+    /// </summary>
+    /// <remarks>
+    /// Lifecycle management is returned to the subscription
+    /// </remarks>
+    /// <param name="client">The client to release claim</param>
     public void ReleaseClient(object client)
     {
         ServiceStub s = client as ServiceStub;
@@ -334,6 +509,13 @@ public class ServiceSubscription
         _subscription.ClaimClient(s.rr_innerstub);
     }
 
+    /// <summary>
+    /// Get or set the connect retry delay in milliseconds
+    /// </summary>
+    /// <remarks>
+    /// Default connect retry delay is 2.5 seconds
+    /// </remarks>
+    /// <value></value>
     public uint ConnectRetryDelay
     {
         get {
@@ -344,29 +526,105 @@ public class ServiceSubscription
         }
     }
 
+    /// <summary>
+    /// Event listener for when a client connects
+    /// </summary>
     public event Action<ServiceSubscription, ServiceSubscriptionClientID, object> ClientConnected;
+    /// <summary>
+    /// Event listener for when a client disconnects
+    /// </summary>
     public event Action<ServiceSubscription, ServiceSubscriptionClientID, object> ClientDisconnected;
 
+    /// <summary>
+    /// Event listener for when a client connection attempt fails. Use to diagnose connection problems
+    /// </summary>
     public event Action<ServiceSubscription, ServiceSubscriptionClientID, string[], Exception> ClientConnectFailed;
 
+    /// <summary>
+    /// Creates a wire subscription
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Wire subscriptions aggregate the value received from the connected services. It can also act as a
+    /// "reverse broadcaster" to send values to clients. See WireSubscription.
+    /// </para>
+    /// <para>
+    /// The optional service path may be an empty string to use the root object in the service. The first level of the
+    /// service path may be "*" to match any service name. For instance, the service path "*.sub_obj" will match
+    /// any service name, and use the "sub_obj" objref.
+    /// </para>
+    /// </remarks>
+    /// <param name="wire_name">The member name of the wire</param>
+    /// <param name="service_path">The service path of the object owning the wire member. 
+    ///     Leave as empty string for root object</param>
+    /// <typeparam name="T">The type of the wire value. This must be specified since the subscription doesn't 
+    /// know the wire value type</typeparam>
+    /// <returns>The wire subscription</returns>
     public WireSubscription<T> SubscribeWire<T>(string wire_name, string service_path = "")
     {
         var s = _subscription.SubscribeWire(wire_name, service_path);
         return new WireSubscription<T>(s);
     }
 
+    /// <summary>
+    /// Creates a pipe subscription
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// Pipe subscriptions aggregate the packets received from the connected services. It can also act as a
+    /// "reverse broadcaster" to send packets to clients. See PipeSubscription.
+    /// </para>
+    /// <para>
+    /// The optional service path may be an empty string to use the root object in the service. The first level of the
+    /// service path may be "*" to match any service name. For instance, the service path "*.sub_obj" will match
+    /// any service name, and use the "sub_obj" objref.
+    /// </para>
+    /// </remarks>
+    /// <param name="pipe_name">The member name of the pipe</param>
+    /// <param name="service_path">The service path of the object owning the pipe member.
+    ///  Leave as empty string for root object</param>
+    /// <param name="max_backlog">The maximum number of packets to store in receive queue</param>
+    /// <typeparam name="T">The type of the pipe packets. This must be specified since the subscription does not 
+    /// know the pipe packet type</typeparam>
+    /// <returns>The pipe subscription</returns>
     public PipeSubscription<T> SubscribePipe<T>(string pipe_name, string service_path = "", int max_backlog = -1)
     {
         var s = _subscription.SubscribePipe(pipe_name, service_path, max_backlog);
         return new PipeSubscription<T>(s);
     }
 
+    /// <summary>
+    /// Get the "default client" connection
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// The "default client" is the "first" client returned from the connected clients map. This is effectively
+    /// default, and is only useful if only a single client connection is expected. This is normally true
+    /// for RobotRaconteurNode.SubscribeService()
+    /// </para>
+    /// <para>
+    /// Clients using GetDefaultClient() should not store a reference to the client. It should instead
+    /// call GetDefaultClient() right before using the client to make sure the most recenty connection
+    /// is being used. If possible, SubscribePipe() or SubscribeWire() should be used so the lifecycle
+    /// of pipes and wires can be managed automatically.
+    /// </para>
+    /// </remarks>
+    /// <returns>The client connection. Cast to expected object type</returns>
     public object GetDefaultClient()
     {
         var s = _subscription.GetDefaultClient();
         return GetClientStub(s);
     }
 
+    /// <summary>
+    /// Try getting the "default client" connection
+    /// </summary>
+    /// <remarks>
+    /// Same as GetDefaultClient(), but returns a bool success instead of throwing
+    /// exceptions on failure.
+    /// </remarks>
+    /// <param name="obj">[out] The client connection</param>
+    /// <returns>true if client object is valid, false otherwise</returns>
     public bool TryGetDefaultClient(out object obj)
     {
         var res = _subscription.TryGetDefaultClient();
@@ -381,12 +639,40 @@ public class ServiceSubscription
         return true;
     }
 
+    /// <summary>
+    /// Get the "default client" connection, waiting with timeout if not connected
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// The "default client" is the "first" client returned from the connected clients map. This is effectively
+    /// default, and is only useful if only a single client connection is expected. This is normally true
+    /// for RobotRaconteurNode.SubscribeService()
+    /// </para>
+    /// <para>
+    /// Clients using GetDefaultClient() should not store a reference to the client. It should instead
+    /// call GetDefaultClient() right before using the client to make sure the most recently connection
+    /// is being used. If possible, SubscribePipe() or SubscribeWire() should be used so the lifecycle
+    /// of pipes and wires can be managed automatically.
+    /// </para>
+    /// </remarks>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>The client connection. Cast to expected object type</returns>
     public object GetDefaultClientWait(int timeout = -1)
     {
         var s = _subscription.GetDefaultClientWait(timeout);
         return GetClientStub(s);
     }
 
+    /// <summary>
+    /// Try getting the "default client" connection, waiting with timeout if not connected
+    /// </summary>
+    /// <remarks>
+    /// Same as GetDefaultClientWait(), but returns a bool success instead of throwing
+    /// exceptions on failure.
+    /// </remarks>
+    /// <param name="obj">[out] The client connection</param>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>true if client object is valid, false otherwise</returns>
     public bool TryGetDefaultClientWait(out object obj, int timeout = -1)
     {
         var res = _subscription.TryGetDefaultClientWait(timeout);
@@ -401,6 +687,14 @@ public class ServiceSubscription
         return true;
     }
 
+    /// <summary>
+    /// Asynchronously get the default client, with optional timeout
+    /// </summary>
+    /// <remarks>
+    /// Same as GetDefaultClientWait(), but async.
+    /// </remarks>
+    /// <param name="timeout">Timeout in milliseconds, or RR_TIMEOUT_INFINITE for no timeout</param>
+    /// <returns>Task that upon completion returns the client object</returns>
     public async Task<object> AsyncGetDefaultClient(int timeout = -1)
     {
         AsyncStubReturnDirectorImpl<object> h = new AsyncStubReturnDirectorImpl<object>(null);
@@ -409,6 +703,14 @@ public class ServiceSubscription
         return await h.Task;
     }
 
+    /// <summary>
+    /// Get the service connection URL
+    /// </summary>
+    /// <remarks>
+    /// Returns the service connection URL. Only valid when subscription was created using
+    /// RobotRaconteurNode.SubscribeService(). Will throw an exception if subscription
+    /// was opened using RobotRaconteurNode.SubscribeServiceByType()
+    /// </remarks>
     public string[] ServiceURL
     {
         get {
@@ -416,6 +718,19 @@ public class ServiceSubscription
         }
     }
 
+    /// <summary>
+    /// Update the service connection URL
+    /// </summary>
+    /// <remarks>
+    /// Updates the URL used to connect to the service. If close_connected is true,
+    /// existing connections will be closed. If false,
+    /// existing connections will not be closed.
+    /// </remarks>
+    /// <param name="url">The new URL to use to connect to service</param>
+    /// <param name="username">(Optional) The new username</param>
+    /// <param name="credentials">(Optional) The new credentials</param>
+    /// <param name="objecttype">(Optional) The desired root object proxy type. Optional but highly recommended.</param>
+    /// <param name="close_connected">(Optional, default false) Close existing connections</param>
     public void UpdateServiceURL(string[] url, string username = null, Dictionary<string, object> credentials = null,
                                  string objecttype = null, bool close_connected = false)
     {
@@ -444,6 +759,19 @@ public class ServiceSubscription
         }
     }
 
+    /// <summary>
+    /// Update the service connection URL
+    /// </summary>
+    /// <remarks>
+    /// Updates the URL used to connect to the service. If close_connected is true,
+    /// existing connections will be closed. If false,
+    /// existing connections will not be closed.
+    /// </remarks>
+    /// <param name="url">The new URL to use to connect to service</param>
+    /// <param name="username">(Optional) The new username</param>
+    /// <param name="credentials">(Optional) The new credentials</param>
+    /// <param name="objecttype">(Optional) The desired root object proxy type. Optional but highly recommended.</param>
+    /// <param name="close_connected">(Optional, default false) Close existing connections</param>
     public void UpdateServiceURL(string url, string username = null, Dictionary<string, object> credentials = null,
                                  string objecttype = null, bool close_connected = false)
     {
@@ -470,6 +798,34 @@ public class ServiceSubscription
     }
 }
 
+/// <summary>
+/// Subscription for wire members that aggregates the values from client wire connections
+/// </summary>
+/// <remarks>
+/// <para>
+/// Wire subscriptions are created using the ServiceSubscription.SubscribeWire() function. This function takes the
+/// type of the wire value, the name of the wire member, and an optional service path of the service
+/// object that owns the wire member.
+/// </para>
+/// <para>
+/// Wire subscriptions aggregate the InValue from all active wire connections. When a client connects,
+/// the wire subscriptions will automatically create wire connections to the wire member specified
+/// when the WireSubscription was created using ServiceSubscription::SubscribeWire(). The InValue of
+/// all the active wire connections are collected, and the most recent one is used as the current InValue
+/// of the wire subscription. The current value, the timespec, and the wire connection can be accessed
+/// using GetInValue() or TryGetInValue().
+/// </para>
+/// <para>
+/// The lifespan of the InValue can be configured using SetInValueLifespan(). It is recommended that
+/// the lifespan be configured, so that the value will expire if the subscription stops receiving
+/// fresh in values.
+/// </para>
+/// <para>
+/// The wire subscription can also be used to set the OutValue of all active wire connections. This behaves
+/// similar to a "reverse broadcaster", sending the same value to all connected services.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The value type used by the wire</typeparam>
 public class WireSubscription<T>
 {
     class WrappedWireSubscriptionDirectorNET : WrappedWireSubscriptionDirector
@@ -516,6 +872,12 @@ public class WireSubscription<T>
         subscription.SetRRDirector(director, id);
     }
 
+    /// <summary>
+    /// Get the current InValue
+    /// </summary>
+    /// <remarks>
+    /// Throws ValueNotSetException if no valid value is available
+    /// </remarks>
     public T InValue
     {
         get {
@@ -531,6 +893,14 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Get the current InValue and metadata
+    /// </summary>
+    /// <remarks>
+    /// Throws ValueNotSetException if no valid value is available
+    /// </remarks>
+    /// <param name="time">[out] the LastValueReceivedTime of the InValue</param>
+    /// <returns>The current InValue</returns>
     public T GetInValue(out TimeSpec time)
     {
         TimeSpec t = new TimeSpec();
@@ -546,6 +916,15 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Try getting the current InValue
+    /// </summary>
+    /// <remarks>
+    /// Same as InValue, but returns a bool for success or failure instead of throwing
+    /// an exception.
+    /// </remarks>
+    /// <param name="value">[out] the current InValue</param>
+    /// <returns>true if value is valid, otherwise false</returns>
     public bool TryGetInValue(out T value)
     {
         var m = new WrappedService_typed_packet();
@@ -568,6 +947,16 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Try getting the current InValue and metadata
+    /// </summary>
+    /// <remarks>
+    /// Same as GetInValue(), but returns a bool for success or failure instead of throwing
+    /// an exception.
+    /// </remarks>
+    /// <param name="value">[out] the current InValue</param>
+    /// <param name="time">[out] the LastValueReceivedTime of the InValue</param>
+    /// <returns>true if value is valid, otherwise false</returns>
     public bool TryGetInValue(out T value, out TimeSpec time)
     {
         var m = new WrappedService_typed_packet();
@@ -592,11 +981,25 @@ public class WireSubscription<T>
         }
     }
 
+
+    /// <summary>
+    /// Wait for a valid InValue to be received from a client
+    /// </summary>
+    /// <remarks>
+    /// Blocks the current thread until value is received or timeout
+    /// </remarks>
+    /// <param name="timeout">The timeout in milliseconds</param>
+    /// <returns>true if value is valid, otherwise false</returns>
     public bool WaitInValueValid(int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE)
     {
         return _subscription.WaitInValueValid(timeout);
     }
 
+    /// <summary>
+    /// Get or Set if InValue is ignored
+    /// </summary>
+    /// <remarks></remarks>
+    /// <value></value>
     public bool IgnoreInValue
     {
         get {
@@ -607,6 +1010,13 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Get or Set the InValue lifespan in milliseconds
+    /// </summary>
+    /// <remarks>
+    /// Get the lifespan of InValue in milliseconds. The value will expire after the specified
+    /// lifespan, becoming invalid. Use -1 for infinite lifespan.
+    /// </remarks>
     public int InValueLifespan
     {
         get {
@@ -617,6 +1027,14 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Set the OutValue for all active wire connections
+    /// </summary>
+    /// <remarks>
+    /// Behaves like a "reverse broadcaster". Calls WireConnection.SetOutValue()
+    /// for all connected wire connections.
+    /// </remarks>
+    /// <param name="value">The new OutValue</param>
     public void SetOutValueAll(T value)
     {
         var iter = new WrappedWireSubscription_send_iterator(_subscription);
@@ -640,6 +1058,9 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Get the number of wire connections currently connected
+    /// </summary>
     public uint ActiveWireConnectionCount
     {
         get {
@@ -647,6 +1068,13 @@ public class WireSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Closes the wire subscription
+    /// </summary>    
+    /// <remarks>
+    /// Wire subscriptions are automatically closed when the parent ServiceSubscription is closed
+    /// or when the node is shut down.
+    /// </remarks>
     public void Close()
     {
         _subscription.Close();
@@ -655,6 +1083,34 @@ public class WireSubscription<T>
     public event Action<WireSubscription<T>, T, TimeSpec> WireValueChanged;
 }
 
+/// <summary>
+/// Subscription for pipe members that aggregates incoming packets from client pipe endpoints
+/// </summary>
+/// <remarks>
+/// <para>
+/// Pipe subscriptions are created using the ServiceSubscription.SubscribePipe() function. This function takes the
+/// the type of the pipe packets, the name of the pipe member, and an optional service path of the service
+/// object that owns the pipe member.
+/// </para>
+/// <para>
+/// Pipe subscriptions collect all incoming packets from connect pipe endpoints. When a client connects,
+/// the pipe subscription will automatically connect a pipe endpoint the pipe endpoint specified when
+/// the PipeSubscription was created using ServiceSubscription.SubscribePipe(). The packets received
+/// from each of the collected pipes are collected and placed into a common receive queue. This queue
+/// is read using ReceivePacket(), TryReceivePacket(), or TryReceivePacketWait(). The number of packets
+/// available to receive can be checked using Available().
+/// </para>
+/// <para>
+/// Pipe subscriptions can also be used to send packets to all connected pipe endpoints. This is done
+/// with the AsyncSendPacketAll() function. This function behaves somewhat like a "reverse broadcaster",
+/// sending the packets to all connected services.
+/// </para>
+/// <para>
+/// If the pipe subscription is being used to send packets but not receive them, the SetIgnoreInValue()
+/// should be set to true to prevent packets from queueing.
+/// </para>
+/// </remarks>
+/// <typeparam name="T">The type of the pipe packets</typeparam>
 public partial class PipeSubscription<T>
 {
     class WrappedPipeSubscriptionDirectorNET : WrappedPipeSubscriptionDirector
@@ -692,6 +1148,13 @@ public partial class PipeSubscription<T>
         subscription.SetRRDirector(director, id);
     }
 
+    /// <summary>
+    /// Dequeue a packet from the receive queue
+    /// </summary>
+    /// <remarks>
+    /// If the receive queue is empty, an InvalidOperationException() is thrown
+    /// </remarks>
+    /// <returns>The dequeued packet</returns>
     public T ReceivePacket()
     {
         using (var m = _subscription.ReceivePacket())
@@ -703,6 +1166,15 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Try dequeuing a packet from the receive queue
+    /// </summary>
+    /// <remarks>
+    /// Same as ReceivePacket(), but returns a bool for success or failure instead of throwing
+    /// an exception
+    /// </remarks>
+    /// <param name="packet">[out] the dequeued packet</param>
+    /// <returns>true if packet dequeued successfully, otherwise false if queue is empty</returns>
     public bool TryReceivePacket(out T packet)
     {
         using (var m = new WrappedService_typed_packet())
@@ -725,6 +1197,14 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Try dequeuing a packet from the receive queue, optionally waiting or peeking the packet
+    /// </summary>
+    /// <param name="packet">[out] the dequeued packet</param>
+    /// <param name="timeout">The time to wait for a packet to be received in milliseconds if the queue is empty, or
+    /// RR_TIMEOUT_INFINITE to wait forever</param>
+    /// <param name="peek">If true, the packet is returned, but not dequeued. If false, the packet is dequeued</param>
+    /// <returns>true if packet dequeued successfully, otherwise false if queue is empty or timed out</returns>
     public bool TryReceivePacketWait(out T packet, int timeout = RobotRaconteurNode.RR_TIMEOUT_INFINITE,
                                      bool peek = false)
     {
@@ -747,6 +1227,12 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Get the number of packets available to receive
+    /// </summary>
+    /// <remarks>
+    /// Use ReceivePacket(), TryReceivePacket(), or TryReceivePacketWait() to receive the packet
+    /// </remarks>
     public uint Available
     {
         get {
@@ -754,6 +1240,9 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Get or set if incoming packets are ignored
+    /// </summary>
     public bool IgnoreReceived
     {
         get {
@@ -764,6 +1253,14 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Sends a packet to all connected pipe endpoints
+    /// </summary>
+    /// <remarks>
+    /// Calls AsyncSendPacket() on all connected pipe endpoints with the specified value.
+    /// Returns immediately, not waiting for transmission to complete.
+    /// </remarks>
+    /// <param name="value">The packet to send</param>
     public void AsyncSendPacketAll(T value)
     {
         var iter = new WrappedPipeSubscription_send_iterator(_subscription);
@@ -787,6 +1284,9 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Get the number of pipe endpoints currently connected
+    /// </summary>
     public uint ActivePipeEndpointCount
     {
         get {
@@ -794,11 +1294,20 @@ public partial class PipeSubscription<T>
         }
     }
 
+    /// <summary>
+    /// Closes the pipe subscription
+    /// </summary>    
+    /// <remarks>
+    /// Pipe subscriptions are automatically closed when the parent ServiceSubscription is closed
+    /// or when the node is shut down.
+    /// </remarks>
     public void Close()
     {
         _subscription.Close();
     }
-
+    /// <summary>
+    /// Listener event for when a pipe packet is received
+    /// </summary>        
     public event Action<PipeSubscription<T>> PipePacketReceived;
 }
 
@@ -871,6 +1380,18 @@ public partial class RobotRaconteurNode
         return filter2;
     }
 
+    /// <summary>
+    /// Subscribe to listen for available services information
+    /// </summary>
+    /// <remarks>
+    /// A ServiceInfo2Subscription will track the availability of service types and
+    /// inform when services become available or are lost. If connections to
+    /// available services are also required, ServiceSubscription should be used.
+    /// </remarks>
+    /// <param name="service_types">An array of service types to listen for, ie 
+    ///     `com.robotraconteur.robotics.robot.Robot`</param>
+    /// <param name="filter">A filter to select individual services based on specified criteria</param>
+    /// <returns>The active subscription</returns>
     public ServiceInfo2Subscription SubscribeServiceInfo2(string[] service_types,
                                                           ServiceSubscriptionFilter filter = null)
     {
@@ -884,6 +1405,17 @@ public partial class RobotRaconteurNode
         return new ServiceInfo2Subscription(sub1);
     }
 
+    /// <summary>
+    /// Subscribe to listen for available services and automatically connect
+    /// </summary>
+    /// <remarks>
+    /// A ServiceSubscription will track the availability of service types and
+    /// create connections when available.
+    /// </remarks>
+    /// <param name="service_types">An arrayof service types to listen for, ie 
+    ///     `com.robotraconteur.robotics.robot.Robot`</param>
+    /// <param name="filter">A filter to select individual services based on specified criteria</param>
+    /// <returns>The active subscription</returns>
     public ServiceSubscription SubscribeServiceByType(string[] service_types, ServiceSubscriptionFilter filter = null)
     {
         var filter2 = SubscribeService_LoadFilter(filter);
@@ -896,12 +1428,36 @@ public partial class RobotRaconteurNode
         return new ServiceSubscription(sub1);
     }
 
+    /// <summary>
+    /// Subscribe to a service using one a URL. Used to create robust connections to services
+    /// </summary>
+    /// <remarks>
+    /// Creates a ServiceSubscription assigned to a service with one or more candidate connection URLs. The
+    /// subscription will attempt to maintain a persistent connection, reconnecting if the connection is lost.
+    /// </remarks>
+    /// <param name="url">One or more candidate connection urls</param>
+    /// <param name="username">An optional username for authentication</param>
+    /// <param name="credentials">Optional credentials for authentication</param>
+    /// <param name="objecttype">The desired root object proxy type. Optional but highly recommended.</param>
+    /// <returns>The active subscription</returns>
     public ServiceSubscription SubscribeService(string url, string username = null,
                                                 Dictionary<string, object> credentials = null, string objecttype = null)
     {
         return SubscribeService(new string[] { url }, username, credentials, objecttype);
     }
 
+    /// <summary>
+    /// Subscribe to a service using one or more URL. Used to create robust connections to services
+    /// </summary>
+    /// <remarks>
+    /// Creates a ServiceSubscription assigned to a service with one or more candidate connection URLs. The
+    /// subscription will attempt to maintain a persistent connection, reconnecting if the connection is lost.
+    /// </remarks>
+    /// <param name="url">One or more candidate connection urls</param>
+    /// <param name="username">An optional username for authentication</param>
+    /// <param name="credentials">Optional credentials for authentication</param>
+    /// <param name="objecttype">The desired root object proxy type. Optional but highly recommended.</param>
+    /// <returns>The active subscription</returns>
     public ServiceSubscription SubscribeService(string[] url, string username = null,
                                                 Dictionary<string, object> credentials = null, string objecttype = null)
     {
