+++ RobotRaconteurNET/RobotRaconteurNET/Util.cs
@@ -40,6 +40,9 @@ class WrappedRobotRaconteurExceptionHelper : RobotRaconteurExceptionHelper
     }
 }
 
+/// <summary>
+/// Custom LogRecordHandler that sends log records to a callback function
+/// </summary>
 public class UserLogRecordHandler : UserLogRecordHandlerBase
 {
     class UserLogRecordHandlerDirectorNET : UserLogRecordHandlerDirector
@@ -55,11 +58,86 @@ public class UserLogRecordHandler : UserLogRecordHandlerBase
             handler?.Invoke(record);
         }
     }
+
+    /// <summary>
+    /// Construct a UserLogRecordHandler with a specified callback. Use RobotRaconteurNode.SetRecordLogRecordHandler()
+    /// to set the record handler.
+    /// </summary>
+    /// <param name="handler">The callback to receive log records</param>
     public UserLogRecordHandler(Action<RRLogRecord> handler)
     {
         UserLogRecordHandlerDirectorNET director = new UserLogRecordHandlerDirectorNET(handler);
         int id = RRObjectHeap.AddObject(director);
         _SetHandler(director, id);
+    }
+}
+
+public class RRObjectHeap
+{
+
+    class RRObjectHeap_support : RRNativeObjectHeapSupport
+    {
+        public override void DeleteObject_i(int id)
+        {
+            RRObjectHeap.DeleteObject(id);
+        }
+    }
+
+    static RRObjectHeap_support support;
+
+    static RRObjectHeap()
+    {
+        support = new RRObjectHeap_support();
+        RRNativeObjectHeapSupport.Set_Support(support);
+    }
+
+    public static Dictionary<int, object> objects = new Dictionary<int, object>();
+    public static int objectcount = 0;
+
+    public static int AddObject(object o)
+    {
+        lock (objects)
+        {
+            if (objectcount > int.MaxValue - 10)
+                objectcount = 0;
+            objectcount++;
+            while (objects.ContainsKey(objectcount))
+                objectcount++;
+
+            int i = objectcount;
+            objects.Add(i, o);
+            // Console.WriteLine(objects.Count);
+            return i;
+        }
+    }
+
+    public static void DeleteObject(int id)
+    {
+        lock (objects)
+        {
+            try
+            {
+                objects.Remove(id);
+                // Console.WriteLine(objects.Count);
+            }
+            catch
+            {
+
+                // Console.WriteLine(e.ToString());
+            }
+        }
+    }
+
+    public static object GetObject(int id)
+    {
+        try
+        {
+            return objects[id];
+        }
+        catch (KeyNotFoundException)
+        {
+            throw new ApplicationException("Internal stub error");
+        }
     }
 }
